% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-pulsed-interactive.R
\name{acq_fluence_interactive}
\alias{acq_fluence_interactive}
\alias{acq_fraction_pulsed_interactive}
\title{Acquire flash spectra interactively}
\usage{
acq_fluence_interactive(integ.time = 2, num.scans = 1L,
  num.exposures = 1L, f.trigger.pulses = f.trigger.message,
  protocols = NULL, correction.method = NA, descriptors = NA,
  stray.light.method = "none", qty.out = "fluence", save.pdfs = TRUE,
  interface.mode = "manual")

acq_fraction_pulsed_interactive(integ.time = 2, num.scans = 1L,
  num.exposures = 1L, f.trigger.pulses = f.trigger.message,
  protocols = NULL, correction.method = NA, descriptors = NA,
  ref.value = 1, ref.absolute = TRUE, ref.persistence = 1,
  qty.out = "Tfr", type = "total", stray.light.method = "simple",
  save.pdfs = TRUE, interface.mode = "manual")
}
\arguments{
\item{integ.time}{numeric Integration time for a measurement event
in seconds.}

\item{num.scans}{integer Number of scans to average.}

\item{num.exposures}{integer Number or light pulses (flashes) per scan. Set
to \code{-1L} to indicate that the light source is continuous.}

\item{f.trigger.pulses}{function Function to be called to trigger light
pulse(s). Should accept as its only argument the number of pulses, and
return \code{TRUE} on sucess and \code{FALSE} on failure.}

\item{protocols}{list of character vectors.}

\item{correction.method}{list The method to use when applying the calibration}

\item{descriptors}{list A list of instrument descriptors containing
calibration data.}

\item{stray.light.method}{character Used only when the correction method is
created on-the-fly.}

\item{qty.out}{character One of "Tfr" (spectral transmittance as a fraction of one),
"fluence" (spectral irardiance), "cps" (counts per second), or "raw" (raw sensor counts).}

\item{save.pdfs}{logical Whether to save PDFs to files or not.}

\item{interface.mode}{character One of "manual", and "manual-attr".}

\item{ref.value}{numeric or filter_spct/reflector_spct object.}

\item{ref.absolute}{logical Flag indicating whether the reference is
absolute or relative. In the last case, cps spectra are normalised.}

\item{ref.persistence}{integer Number of measurements using the same
reference reading.}

\item{type}{character Type of transmittance or reflectance measured.}

\item{qty.out}{character One of "Tfr" (transmittance as a fraction of one)
or "raw" (raw counts).}
}
\description{
Functions providing a simple interactive front-end to the functions in the
package. Mostly as example code that can be modified for diferent
uses.
}
\details{
These functions can be useful as is for commonly done measurements
  but they also play the role of examples that users can modify according
  to their needs. They are all composed in a modular way from functions
  that can be reshuffled and combined with other functions to define
  new variations better suited to users needs and tastes. The examples we
  provide here cover the measurement of spectral irradiance of pulsed light
  sources: spectral fluence from flashes, and transmittance, reflectance
  and absorbance using a flash as light sources.

  The default behaviour of the functions can be changed by passing different
  arguments through parameters, but for special use cases it could be best
  for users to define case specific data acquisition functions from the
  same building blocks.

The different interface modes are suitable for different types of
  measurements.
}
\note{
The integration time is set manually and the flash is triggered within
  this "window" one or more times. The default protocols are usually
  suitable, if new protocols are passed, each character vector must contain
  strings "light", "filter" and "dark", or "sample", "reference", and "dark",
  depending on the function.

The calculations for reflectance and transmittance are very similar,
  so we provide a single function capable of handling both. For
  transmittance the reference is usually direct exposure to radiation but
  for reflectance a white reference patch is normally used. In some cases
  one may want to use a grey reference. We provide an argument that allows
  the user to supply a constant or a spectrum describing the properties of
  the reference. It is also important to distinguish between total and
  internal transmittance, and between total and specular reflectance.
  In both cases which of these is measured depends on the measuring protocol
  (condition used as reference, use of an integrating sphere versus use of a
  probe with a narrow angle of aperture, etc.) and consequently the correct
  value should be entered to ensure that data are correctly tagged and
  later computations valid.
}
\examples{

\dontrun{
# requires an Ocean Optics spectrometer to be connected via USB
acq_fluence_interactive()
}

}
