% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-R.R
\name{check_sn_match}
\alias{check_sn_match}
\title{Check consistency of serial number}
\usage{
check_sn_match(x, correction.method, missmatch.action = stop)
}
\arguments{
\item{x}{a generic_mspct object, in normal use a raw_mspct object.}

\item{correction.method}{a list describing the correction method. Must
incluse a member named \code{"spectrometer.sn"}.}

\item{missmatch.action}{a function, in practice one of \code{stop},
\code{warning}, \code{message} or \code{NULL}.}
}
\value{
A logical vector of length one, with \code{FALSE} indicating a
  missmatch.
}
\description{
This method checks that the sn of the instrument used to acquire the raw
counts matches that in the definition of a correction method.
}
\seealso{
Other spectral data-processing functions:
\code{\link[=MAYP112785_tail_correction]{MAYP112785_tail_correction()}},
\code{\link[=MAYP11278_tail_correction]{MAYP11278_tail_correction()}},
\code{\link[=linearize_counts]{linearize_counts()}},
\code{\link[=merge_raw_mspct]{merge_raw_mspct()}},
\code{\link[=new_correction_method]{new_correction_method()}},
\code{\link[=ref_correction]{ref_correction()}},
\code{\link[=trim_counts]{trim_counts()}},
\code{\link[=uvb_corrections]{uvb_corrections()}}
}
\concept{spectral data-processing functions}
