% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute-calibration.R
\name{compute_irrad_calibration}
\alias{compute_irrad_calibration}
\title{Compute calibration multipliers.}
\usage{
compute_irrad_calibration(FEL.raw.counts, D2.raw.counts,
  pix.wavelengths = NULL, wl.range = c(250, 900), FEL.k, D2.k, method,
  verbose = getOption("photobiology.verbose", default = FALSE))
}
\arguments{
\item{FEL.raw.counts}{source_mscpt The raw counts and integration time data
for the FEL light source.}

\item{D2.raw.counts}{source_mscpt The raw counts and integration time data
for the D2 light source.}

\item{pix.wavelengths}{numeric vector The "true" wavelengths in nanometres at
each pixel in the detector array.}

\item{wl.range}{numeric vector of length two Range of wavelengths for which
to compute the calibration.}

\item{FEL.k}{numeric vector a numeric vector with n constants for the
polynomial for the FEL lamp.}

\item{D2.k}{numeric vector a numeric vector with n constants for the
polynomial for the D2 lamp.}

\item{method}{list Method variant to be used.}

\item{verbose}{Logical indicating the level of warnings wanted. Defaults to
\code{FALSE}.}
}
\value{
An instrument descriptor as a list containing the updated wavelengths
  and multipliers from the calibration.
}
\description{
Function to calculate from a set of three (or optionally five or two) output
files from the spectrometer, as produce by SpectraSuite, for each of two
calibration lamps (D2 and FEL), and calibration data for the lamps, a set of
calibration multipliers for the instrument. This function applies a
non-linearity correction, an slit-function correction, and stray-light
correction, and optionally smoothing by means of process_maya_arrays().
Optionaly a HDR ("high dynamic range") merging of spectra (based on
integration time bracketing) using two different integration times is done.
}
\note{
To calculate new multipliers we set the calibration multipliers equal
  to 1. Do the calculations as usual and calculate new multipliers based on
  the known spectral irradiance for the calibration lamps.
}
