% Not generated by roxygen2: edit by hand
% Please edit this file in man/example-scripts.Rd
\name{example-scripts}
\alias{data-acquisition-scripts}
\title{Data acquisition scripts}
\description{
Package 'ooacquire' includes example scripts for acquisition of spectral data
using Ocean Optics spectrometers. These scripts demonstrate the use of
functions \code{\link{acq_irrad_interactive}()} and
\code{\link{acq_fraction_interactive}()}.
}
\details{
The default behaviour of functions \code{\link{acq_irrad_interactive}()} and
\code{\link{acq_fraction_interactive}()} is controlled by passing arguments
to their formal parameters when called as well as by R options set before
calling them.

Some of these defaults affect the type of data returned and the
computation of physical quantities from the detector raw-counts data received
from the spectrometer and which type of graphical and textual summaries are
computed an saved.

One parameter can be passed as argument the definition
of an R function used to controll a lamp or some other device by a trigger or
a message to the user.

However, the majority of the options make it possible to chose which menues
and features are accesible interactively through menues. Experience has shown
that in a long data acquisition session, displaying to the operator only
those options relevant to the current protocol or measurement approach reduces
operator stress, decreases the oportunities for mistakes and allows measurements
to be done faster and more consistently. The folder, operator, and file names
used can be also set by arguments.

These functions have what at first sight is an excessive number of formal
parameters. This is true if arguments to these parameters are entered manually
each time spectral measurements are acquired. The idea is that using a
simple R script, users can setup the default behaviour of these functions to
match their needs. Once a script is available, it can be sourced to start
a measurement session.

The scripts included in 'ooacquire' provide several examples of how to
customize the default behaviour of the functions. Some of the files also
include commented-out variations on the code example.
}
\examples{
# find where the example files are installed as part of the package
path <- system.file("example-scripts", package = "ooacquire")

# list the files in the folder
list.files(path)

# read and display one example file
readLines(paste(path, "irrad-acq-interac-simple.R", sep = "/"))
}
