% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-rm-jwrapper.R
\name{files_rm_jwrapper}
\alias{files_rm_jwrapper}
\title{Remove Java wrapper from objects in files}
\usage{
files_rm_jwrapper(
  path = ".",
  pattern = "\\\\.spct\\\\.[Rr]da",
  recursive = TRUE,
  date.limit = NULL,
  trim.descriptor = TRUE,
  save.files = FALSE,
  verbose = getOption("photobiology.verbose", default = FALSE)
)
}
\arguments{
\item{path}{character Vector of full path names; the default corresponds to
the working directory. Tilde expansion (see \code{\link{path.expand}()}) is
performed. See \code{\link{list.files}()} for details.}

\item{pattern}{character An optional regular expression. Only file names
which match the regular expression will be returned. See
\code{\link{list.files}()} for details.}

\item{recursive}{logical Should the listing recurse into directories?}

\item{date.limit}{POSIXct Files with more recent creation dates are skipped
assuming they have been created with 'ooacquire' (>= 0.2.5 2022-09-30).}

\item{trim.descriptor}{logical If \code{TRUE} apply
\code{trimInstrDesc()} instead of \code{rm_jwrapper()} to spectral objects
containing data in physical units. \code{rm_jwrapper()} is always used for
objects containing sensor counts, either raw of cps.}

\item{save.files}{logical If \code{FALSE} skip saving of files. Useful to
test what files would be modified.}

\item{verbose}{logical Issue messages when unsupported objects or spectral
objects already free of the problematic Java wrapper are encountered.}
}
\value{
A \code{character} vector with the paths to the updated files,
returned silently. As a side effect files can be saved containing updated
objects while keeping original files renamed as backups. By default no
files are saved.
}
\description{
Remove Java wrappers stored in field \code{"w"} from the
\code{instr.descriptor} metadata attribute of spectral objects and of members
of collections of spectra loaded from R data files, saving them using their
original name and location, after renaming the original files to serve as
backups.
}
\details{
Spectral data objects acquired directly from a connected spectrometer
with 'ooacquire' (< 0.2.5) contain a leftover Java wrapper used to connect to
the spectrometer. This is of no use after disconnection, but if it remains
in the objects saved to an \code{.rda} file loading these files requires
'rJava' and the Java JDK to be available. Some old \code{source_spct} and
\code{filter_spct} objects can also contain the full instrument descriptor
including calibration data, which makes the objects unnecessarily large.

This function, searches folders for files with names matching \code{pattern},
loads them one by one if modified before the \code{date.limit}, removes the
spurious Java wrapper and saves the updated objects into a file with the
original name after renaming the original file.

The defaults match the defaults of \code{s_irrad_corected()} and
\code{acq_irrad_interactive()} in the current version of 'ooacquire'.
Field \code{w} containing the \code{jwrapper} is deleted from all
spectral objects, and for objects containing spectral data expressed
in physical units, also fields related to instrument calibration are
deleted. If \code{trim.descriptor = FALSE} is passed in the call,
deleting the \code{w} is the action applied to all objects.
Any objects in the loaded files that are not spectral objects or
collections of spectral objects are kept unchanged.
}
\section{Date related uncertainties}{

The date-based file selection uses file modification time as Windows has
an unusual logic for file creation time, resulting in creation times being
later than modification time when a file is copied. When attached files are
downloaded or when files are downloaded they may get the creation and
modification times set to the time of downloading. This is only a speed
up for avoiding loading of files. Files containing objects that do not need
updating will not be modified even if they are loaded.
}

\section{Warning!}{
 It is strongly recommended that before committing
changes a test is first done to check which files will be updated.
}

\seealso{
\code{\link{ls}()}, and \code{\link{list.files}()}, which are used
  to list objects and files, and \code{\link{rm_jwrapper}()} used to remove
  the Java wrapper from the instrument descriptor embedded in objects of
  class \code{generic_spct} and derived classes, respectively.

Use function \code{\link{rm_jwrapper}()} for individual
  objects in the R workspace.

Other Functions and methods related to instrument descriptors.:
\code{\link[=files_s_irrad_update]{files_s_irrad_update()}},
\code{\link[=instr_defaults]{instr_defaults()}},
\code{\link[=rm_jwrapper]{rm_jwrapper()}},
\code{\link[=update_bad_pixs]{update_bad_pixs()}}
}
\concept{Functions and methods related to instrument descriptors.}
