% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files-s-irrad-updt.R
\name{files_s_irrad_updt}
\alias{files_s_irrad_updt}
\title{Recompute spectral irradiance in objects in files}
\usage{
files_s_irrad_updt(
  path = ".",
  pattern = "\\\\.spct\\\\.[Rr]da",
  recursive = TRUE,
  date.limit = NULL,
  spct.names = c(light = "light", filter = "filter", dark = "dark"),
  correction.method = NULL,
  hdr.tolerance = getOption("ooacquire.hdr.tolerance", default = 0.05),
  return.cps = NULL,
  trim.descriptor = TRUE,
  which = NULL,
  which.not = NULL,
  save.files = FALSE,
  verbose = getOption("photobiology.verbose", default = FALSE)
)
}
\arguments{
\item{path}{character Vector of full path names; the default corresponds to
the working directory. Tilde expansion (see \code{\link{path.expand}()}) is
performed. See \code{\link{list.files}()} for details.}

\item{pattern}{character An optional regular expression. Only file names
which match the regular expression will be returned. See
\code{\link{list.files}()} for details.}

\item{recursive}{logical Should the listing recurse into directories?}

\item{date.limit}{POSIXct Files with more recent creation dates are skipped
assuming they have been created with 'ooacquire' (>= 0.2.5 2022-09-30).}

\item{spct.names}{named character vector of length three, to map names in
\code{x} to those expected.}

\item{correction.method}{A named list of constants and functions defining the
method to be used for stray light and dark signal corrections.}

\item{hdr.tolerance}{numeric Tolerance for mean deviation among cps columns as
a fraction of one. Used in check of HDR consistency. A negative value
disables merging using only the data for the shortest integration time.}

\item{return.cps}{logical Useful when there is no need to apply a calibration,
such as when computing new calibration multipliers.}

\item{trim.descriptor}{logical Passed to
\code{\link{s_irrad_corrected_updt}()}.  If TRUE the spectrometer
calibration constants, pixel wavelengths, slit-function "tail-correction"
function code and other calibration-related information is deleted.}

\item{which, which.not}{character Names of attributes to copy or
not to copy.}

\item{save.files}{logical If \code{FALSE} skip saving of files. Useful to
test what files would be modified.}

\item{verbose}{logical Issue messages when unsupported objects or spectral
objects already free of the problematic Java wrapper are encountered.}
}
\value{
A \code{character} vector with the paths to the updated files,
returned silently. As a side effect files can be saved containing updated
objects while keeping original files renamed as backups. By default no
files are saved.
}
\description{
Recompute spectral irradiance in \code{source_spct} objects stored in
\code{.rda} files created with \code{\link{acq_irrad_interactive}()} from
earlier versions of 'ooacquire' using the calibrations and methods in the
current version of 'ooacquire'.
}
\details{
Spectral data objects acquired directly from a connected spectrometer
with 'ooacquire' contain embedded calibration data and correction algorithms
that by default are used when spectral irradiance is computed. In addition,
quality control of raw counts data has been enhanced. Function
\code{\link{s_irrad_corrected_updt}()} uses the current matching
descriptor, quality control and conversion default settings.

This function, searches folders for files with names matching \code{pattern},
loads them one by one if modified before the \code{date.limit}, removes the
Java wrapper if present and recomputes spectral irradiance with
\code{\link{s_irrad_corrected_updt}()} and saves the updated objects into a
file with the original name after renaming the original file. If the
\code{"raw_mspct"} object is not accompanied by a \code{"source_spct"}
object with a matching name, the file is skipped.

The defaults match the defaults of \code{s_irrad_corrected()} and
\code{acq_irrad_interactive()} in the current version of 'ooacquire'.
Field \code{w} containing the \code{jwrapper} is deleted from all
spectral objects, and for objects containing spectral data expressed
in physical units, also fields related to instrument calibration are also
deleted by default. If \code{trim.descriptor = FALSE} is passed in the call,
deleting the \code{w} is the action applied to all objects.
}
\section{Date related uncertainties}{

The date-based file selection uses file modification time as Windows has
an unusual logic for file creation time, resulting in creation times being
later than modification time when a file is copied. When attached files are
downloaded or when files are downloaded they may get the creation and
modification times set to the time of downloading. This is only a speed
up for avoiding loading of files. Files containing objects that do not need
updating will not be modified even if they are loaded.
}

\section{Warning!}{
 It is strongly recommended that before committing
changes a test is first done to check which files will be updated.
}

\seealso{
\code{\link{ls}()}, \code{\link{list.files}()},
  \code{\link{s_irrad_corrected_updt}()} which are used to list
  objects and files, and to update spectral irradiance, respectively.

Use function \code{\link{s_irrad_corrected_updt}()} for individual
  objects in the R workspace.

Other Functions and methods related to instrument descriptors.:
\code{\link[=files_rm_jwrapper]{files_rm_jwrapper()}},
\code{\link[=instr_defaults]{instr_defaults()}},
\code{\link[=rm_jwrapper]{rm_jwrapper()}},
\code{\link[=update_bad_pixs]{update_bad_pixs()}}
}
\concept{Functions and methods related to instrument descriptors.}
