% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb-corrections.R
\name{filter_correction}
\alias{filter_correction}
\alias{no_filter_correction}
\title{Correct for stray light}
\usage{
filter_correction(
  x,
  flt,
  stray.light.method = "original",
  stray.light.wl = c(218.5, 228.5),
  flt.dark.wl = c(193, 209.5),
  flt.ref.wl = c(360, 379.5),
  flt.Tfr = 1,
  trim = 0.05,
  filter.nir.adjust = FALSE,
  hdr.tolerance = getOption("ooacquire.hdr.tolerance", default = 0.1),
  verbose = getOption("photobiology.verbose", default = FALSE)
)

no_filter_correction(
  x,
  stray.light.wl = c(218.5, 228.5),
  flt.dark.wl = c(193, 209.5),
  flt.ref.wl = NULL,
  flt.Tfr = 1,
  trim = 0,
  hdr.tolerance = getOption("ooacquire.hdr.tolerance", default = 0.1),
  verbose = getOption("photobiology.verbose", default = FALSE)
)
}
\arguments{
\item{x, flt}{cps_spct objects, containing spectral data from which to subtract
stray light, and measured stray light, respectively.}

\item{stray.light.method}{Method variant used, "original" (Ylianttila),
"simple", "full", "sun", "raw", "none".}

\item{stray.light.wl}{numeric vector of length 2 giving the range of
wavelengths to use for the final stray light correction.}

\item{flt.dark.wl, flt.ref.wl}{numeric vectors of length 2 giving the ranges
of wavelengths to use for the "dark" and "illuminated" regions of the array
in the filter correction.}

\item{flt.Tfr}{numeric fractional transmittance of the filter to the source
of stray light, used only for method "simple".}

\item{trim}{a numeric value to be used as argument for mean}

\item{filter.nir.adjust}{logical Flag indicating if the cps in the
"filter" reference spectrum need to be adjust based on NIR region cps
in the "light" spectrum. EXPERIMENTAL!!}

\item{hdr.tolerance}{numeric Tolerance for mean deviation among cps columns
as a fraction of one. Used in check of HDR consistency.}

\item{verbose}{Logical indicating the level of warnings wanted.}
}
\description{
Correct cps readings for stray light, using either measured stray light, or
using a non-excited region of the detector array.
}
