% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s-irrad-corrected.R
\name{find_spct_names}
\alias{find_spct_names}
\title{Find names of raw-counts members}
\usage{
find_spct_names(
  x,
  qty = "s.irrad",
  verbose = getOption("photobiology.verbose", default = FALSE)
)
}
\arguments{
\item{x}{raw_spct or list object with named members.}

\item{qty}{character The physical quantity whose calculation the spectra in
\code{x} are intended for. Currently only \code{"s.irrad"} is supported.}

\item{verbose}{logical If \code{TRUE} issue informative messages.}
}
\value{
A named list of character vectors containing member names for each
  stage of a measurement protocol.
}
\description{
Find member names matching the different stages of a measurement protocol.
}
\details{
When \code{qty = "irrad"} is passed in the call, search for members
  of \code{x} with names containing each of the character strings
  \code{"dark"}, \code{"filter"}, and \code{"light"}, returning them in a
  list of character vectors with members named \code{"dark"},
  \code{"filter"}, and \code{"light"}. In case of no matching names in
  \code{x} the corresponding member of the returned list is missing. Thus,
  when no matches are found for any of the three strings, the object returned
  is \code{list()}, the empty list.

  When data have been acquired with `acq_irrad_interactive()` both single
  and time series measurements and all default protocols are supported when
  using default measurement protocols or if user created protocols use names
  that contain suitable matched strings embedded.

  When data are imported with `read_files2mspct()` the names are always set
  by the user, and for this function to work correctly, the names used must
  contain suitable matched strings embedded.

  When \code{x} contains members named not following the convention expected
  by \code{find_spct_names()}, a mapping of member names to protocol steps
  can be created using a different approach, either manually or in user
  written code.
}
\examples{
find_spct_names(x = white_grow_LED.raw_mspct)
find_spct_names(x = white_grow_LED.raw_mspct, verbose = TRUE)
find_spct_names(x = sun001.raw_mspct, verbose = TRUE)

}
\seealso{
Function \code{find_spct_names()} is used to set the default
  argument for formal parameter \code{spct.names} in
  \code{\link{s_irrad_corrected}()} and \code{\link{s_irrad_update}()}.
}
