% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame-s-data.R
\docType{data}
\name{FLMS00440_calib_dates.df}
\alias{FLMS00440_calib_dates.df}
\title{Dates for which specific calibrations are valid.}
\format{
A list.
}
\usage{
FLMS00440_calib_dates.df
}
\description{
Data frame with data used to match raw spectral data to different
calibrations based on the date data was acquired.
}
\keyword{datasets}
