% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame-s-data.R
\docType{data}
\name{FLMS00673_none.mthd}
\alias{FLMS00673_none.mthd}
\title{Correction method definition}
\format{A list.}
\usage{
FLMS00673_none.mthd
}
\description{
This a list with data needed to process raw data in the example files. It
contains constants and functions defining a default method with no special
corrections or data processing.
}
\keyword{datasets}
