% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame-s-data.R
\docType{data}
\name{FLMS04133_cal.spct}
\alias{FLMS04133_cal.spct}
\title{Instrument calibration for a Flame S spectrometer}
\format{A \code{calibration_spct} object.}
\usage{
FLMS04133_cal.spct
}
\description{
This a list of instrument descriptors, each one containing the calibration
data needed to read raw data acquired on different dates.
}
\keyword{datasets}
