% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame-s-data.R
\docType{data}
\name{FLMS04133_calib_dates.df}
\alias{FLMS04133_calib_dates.df}
\title{Dates for which the different descriptors conatin valid calibrations.}
\format{A list.}
\usage{
FLMS04133_calib_dates.df
}
\description{
This a data frame with data needed to process raw data acquired on different
dates.
}
\keyword{datasets}
