% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame-s-data.R
\docType{data}
\name{FLMS04133_descriptors}
\alias{FLMS04133_descriptors}
\alias{FLMS04133_cal.spct}
\alias{FLMS04133_calib_dates.df}
\alias{FLMS04133_ylianttila.mthd}
\alias{FLMS04133_sun.mthd}
\alias{FLMS04133_simple.mthd}
\alias{FLMS04133_none.mthd}
\title{Flame S spectrometer s/n FLMS04133}
\format{
Lists and data frames.

An object of class \code{calibration_spct} (inherits from \code{generic_spct}, \code{spec_tbl_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2048 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 6 columns.

An object of class \code{list} of length 0.

An object of class \code{list} of length 0.

An object of class \code{list} of length 11.

An object of class \code{list} of length 11.
}
\usage{
FLMS04133_descriptors

FLMS04133_cal.spct

FLMS04133_calib_dates.df

FLMS04133_ylianttila.mthd

FLMS04133_sun.mthd

FLMS04133_simple.mthd

FLMS04133_none.mthd
}
\description{
Calibration data, descriptors and methods for correction of stray light and
high dynamic range based on multiple integrations times per spectrum.
}
\note{
Some of the method definitions are fill-ins for methods not
  implemented. In this case no characterizations of the slit function are
  available.
}
\seealso{
Other objects containing instrument-specific data:
\code{\link{FLMS00416_descriptors}},
\code{\link{FLMS00440_descriptors}},
\code{\link{FLMS00673_descriptors}},
\code{\link{JAZA3098_calib_dates.df}},
\code{\link{MAYP112785_descriptors}},
\code{\link{MAYP11278_descriptors}},
\code{\link{MAYP114590_descriptors}}
}
\concept{objects containing instrument-specific data}
\keyword{datasets}
