% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame-s-data.R
\docType{data}
\name{FLMS04133_simple.mthd}
\alias{FLMS04133_simple.mthd}
\title{Correction method definition}
\format{A list.}
\usage{
FLMS04133_simple.mthd
}
\description{
This a list with data needed to process raw data in the example files. It
contains constants and functions defining the method developed by Lasse
Ylianttila.
}
\note{
NOT IMPLEMENTED.
}
\keyword{datasets}
