% Not generated by roxygen2: edit by hand
% Please edit this file in man/foreign-files.Rd
\name{foreign-files}
\alias{foreign-data-files}
\title{Example foreign data files}
\description{
Package 'ooacquire' includes example data files. Some of these files are in
\emph{foreign} formats, mostly variations of text file formats with raw-counts
data obtained with the software from spectrometer manufacturer Ocean Optics.
}
\details{
These files are used in the documentation and for testing file import functions.
The files can be also useful when learning to use 'ooacquire'. The collection
is incomplete as it covers only a few models of spectrometers. There is some
variation in the file format used by Ocean Optics. \strong{Please share your
files acquired with Ocean Optics spectrometers, especially examples of those
that fail to be correctly imported, if you encounter them.}
}
\seealso{Package \code{\link[photobiologyInOut]{photobiologyInOut}}
supports the import of additional formats, including those acquired with
other brands of spectrometers. However, the conversion of raw counts into
physical quantities is not directly supported for these other spectrometers.
}
\examples{
# find where the foreign files are installed as part of the package
path <- system.file("extdata", "foreign-files", package = "ooacquire")

# list the files in the folder
list.files(path)

# read the top 20 lines of an example file
readLines(paste(path, "led-light.txt", sep = "/"), n = 20)
}
