% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{format_idx}
\alias{format_idx}
\title{Format index with enough leading zeros}
\usage{
format_idx(idx, max.idx = NULL)
}
\arguments{
\item{idx}{integer vector Numbers to convert into character strings with
enough leading zeros to accommodate \code{max.idx}.}

\item{max.idx}{integer vector of length one.}
}
\description{
Format index with enough leading zeros
}
