% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{get_oo_descriptor}
\alias{get_oo_descriptor}
\title{Get the instrument description and EEPROM data}
\usage{
get_oo_descriptor(
  w,
  sr.index = 0L,
  ch.index = 0L,
  area = NULL,
  diff.type = NULL
)
}
\arguments{
\item{w}{an open Wrapper object from OmniDriver.}

\item{sr.index}{an index to address the spectrometer for the time being not
exported.}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}

\item{area}{numeric Passed to \code{o_calib2irrad_mult()}.}

\item{diff.type}{character Passed to \code{o_calib2irrad_mult()}.}
}
\value{
a list
}
\description{
Model, configuration, serial number, and calibration data stored in the
EEPROM of an Ocean Optics spectrometer are retrieved and returned in a
list.
}
\details{
Instrument descriptors are \code{"list"} objects with fields
  describing features of a spectrometer that remain consistent across
  data acquisition events, including calibration data, entrance optics,
  serial number, optical configuration, etc. Some of these field can be
  obtained from the EEPROM inside Ocean Optics spectrometers.

  One and only one of \code{area} or \code{diff.type} is needed if an
  irradiance calibration stored in the EEPROM is to be retrieved. If both
  are null, as by default, the irradiance calibration factors will not be
  retrieved even if present in the EEPROM.
}
\seealso{
Other Functions and methods related to instrument descriptors:
\code{\link[=set_descriptor_bad_pixs]{set_descriptor_bad_pixs()}},
\code{\link[=set_descriptor_entrance_optics]{set_descriptor_entrance_optics()}},
\code{\link[=set_descriptor_integ_time]{set_descriptor_integ_time()}},
\code{\link[=set_descriptor_irrad_mult]{set_descriptor_irrad_mult()}},
\code{\link[=set_descriptor_nl]{set_descriptor_nl()}},
\code{\link[=set_descriptor_wl]{set_descriptor_wl()}}
}
\concept{Functions and methods related to instrument descriptors}
