% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{get_oo_settings}
\alias{get_oo_settings}
\title{Get the current values of instrument settings}
\usage{
get_oo_settings(descriptor)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}
}
\value{
a list
}
\description{
Query the spectrometer for the settings currently in use for corrections,
smoothing and acquisition parameters integration time and number of scans.
}
\details{
Instrument settings, are settings whose values can vary between data
  acquisition events. They are stored as a list. The descriptor object is used
  to access the spectrometer and the currently in use settings are retrieved,
  formated and returned.
}
\concept{Functions and methods for instrument settings}
