% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs-acq-raw.R
\name{hs_acq_raw_spct}
\alias{hs_acq_raw_spct}
\title{Measure spectra at high speed}
\usage{
hs_acq_raw_spct(
  descriptor,
  acq.settings,
  num.spectra = 100L,
  base.name = "light",
  f.trigger.pulses = f.trigger.message,
  what.measured = NA,
  where.measured = data.frame(lon = NA_real_, lat = NA_real_),
  set.all = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}.}

\item{acq.settings}{list as returned by functions \code{tune_acq_settings}.}

\item{num.spectra}{integer Number of individual spectra to acquire.}

\item{base.name}{character The name given to individual spectra is formed
by this string followed by a sequential numeric index.}

\item{f.trigger.pulses}{function Function to be called to trigger an
action. Should accept as its only argument the number of pulses, and
return \code{TRUE} on sucess and \code{FALSE} on failure.}

\item{what.measured}{value used to set attribute.}

\item{where.measured}{data.frame with at least columns "lon" and "lat"
compatible with value returned by \code{ggmap::geocode()}.}

\item{set.all}{logical resend or not all instrument settings.}

\item{verbose}{logical to enable or disable warnings.}
}
\value{
A \code{raw_mspct} containing one \code{raw_spct} object with one
  column \code{w.length} and one column \code{counts} for each spectrum. The
  number of columns with raw counts is is always one and integration time
  bracketing or HDR values in \code{acq.settings} are ignored except for the
  smallest value.
}
\description{
Take one set of spectra using the same instruments settings at high speed
using the special OmniDriver API functions. No HDR bracketing is possible,
and synchorization with a pulsed light source is not supported.
}
\seealso{
For normal speed acquisition of a single spectrum and multiple
  spectra according to a user defined protocol see \code{\link{acq_raw_spct}}
  and \code{\link{acq_raw_mspct}}.

Other raw-counts-spectra acquisition functions: 
\code{\link{acq_raw_mspct}()},
\code{\link{acq_raw_spct}()}
}
\concept{raw-counts-spectra acquisition functions}
