% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-irrad-interactive.R
\name{irrad_summary_table}
\alias{irrad_summary_table}
\title{Summarize spectral irradiance or fluence}
\usage{
irrad_summary_table(
  mspct,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  scale.factor = ifelse(unit.out == "photon", 1e+06, 1),
  attr2tb = "when.measured",
  summary.type = "plant",
  digits = 3L
)
}
\arguments{
\item{mspct}{A source_mspct, or a source_spct object containing spectral
irradiance for one or more sources.}

\item{unit.out}{character One of "photon" or "energy".}

\item{scale.factor}{numeric A multiplicative factor used to rescale data.}

\item{attr2tb}{character Vector with one or more of "when.measured",
"what.measured", "where.measured", "how.measured" and "comment".}

\item{summary.type}{character One of "plant", "PAR" or "VIS".}

\item{digits}{integer The number of significant digits in the output.}
}
\value{
A tibble with one row per spectrum and one column per
   summary quantity and attribute and a column with the names of the spectra.
}
\description{
Compute irradiance or fluence by waveband and energy or photon ratios
between wavebands of interest to plants' and human visual responses to light.
}
\details{
This function packages different functions from pacakge 'photobiology'
   and returns a typical set of summaries for different purposes.
}
\examples{

irrad_summary_table(sun.spct)
irrad_summary_table(sun.spct, attr2tb = c("what.measured", "where.measured"))
irrad_summary_table(sun.spct, summary.type = "plant", unit.out = "photon")
irrad_summary_table(sun.spct, summary.type = "PAR", unit.out = "photon")
irrad_summary_table(sun.spct, summary.type = "VIS", unit.out = "energy")

}
\seealso{
See the documentation for functions
  \code{\link[photobiology]{irrad}},
  \code{\link[photobiology]{q_ratio}}, \code{\link[photobiology]{e_ratio}},
  \code{\link[photobiology]{add_attr2tb}},
  \code{\link[photobiologyInOut]{spct_CRI}},
  \code{\link[photobiologyInOut]{spct_CCT}}
  and \code{\link[base]{signif}} which are called to build
  the summary table.
}
