% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaz-data.R
\docType{data}
\name{JAZA3098_calib_dates.df}
\alias{JAZA3098_calib_dates.df}
\title{Dates for which the different descriptors contain valid calibrations.}
\format{A list.}
\usage{
JAZA3098_calib_dates.df
}
\description{
This is a data frame with data needed to process raw data acquired on different
dates.
}
\keyword{datasets}
