% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaz-data.R
\docType{data}
\name{JAZA3098_ch1_descriptors}
\alias{JAZA3098_ch1_descriptors}
\title{Instrument descriptors for a JAZ  spectrometer}
\format{A list.}
\usage{
JAZA3098_ch1_descriptors
}
\description{
This a list of instrument descriptors, each one containing the calibration
data needed to read raw data acquired on different dates.
}
\keyword{datasets}
