% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaz-data.R
\docType{data}
\name{JAZA3098_ch2_ylianttila.mthd}
\alias{JAZA3098_ch2_ylianttila.mthd}
\title{Correction method definition}
\format{A list.}
\usage{
JAZA3098_ch2_ylianttila.mthd
}
\description{
This a list with data needed to process raw data in the example files. It
contains constants and functions defining the method developed by Lasse
Ylianttila.
}
\note{
NOT IMPLEMENTED.
}
\keyword{datasets}
