% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearize.count.R
\name{linearize_count}
\alias{linearize_count}
\title{Function to apply linearization correction to raw counts data.}
\usage{
linearize_count(x, linearize.fun = NULL, force_zero = TRUE,
  verbose = FALSE)
}
\arguments{
\item{x}{raw_spct object.}

\item{linearize.fun}{A function or a polynom::polynomial object containing
the linearization to be applied.}

\item{force_zero}{A logical indicating whether to change negative count
values to zero.}

\item{verbose}{Logical Currently ignired.}
}
\value{
A raw_spct object containing the adjusted values, still as uncalibrated
  counts. The object is tagged with the with attribute "linearized" set to
  the function used for linearization.
}
\description{
Uses a user supplied function, possibly that supplied by a manufacturer
like Ocean Optics stored in firmware or in any other form.
}
\author{
Pedro J. Aphalo, based on Excel code by Lasse Ylianttila.
}
\references{
\url{http://www.r4photobiology.info/}
}
\keyword{misc}

