% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearize.counts.R
\name{linearize_counts}
\alias{linearize_counts}
\title{Function to apply linearization correction to raw counts data.}
\usage{
linearize_counts(
  x,
  force.zero = TRUE,
  verbose = getOption("photobiology.verbose", default = FALSE)
)
}
\arguments{
\item{x}{raw_spct object.}

\item{force.zero}{A logical indicating whether to change negative count
values to zero.}

\item{verbose}{Logical Currently ignored.}
}
\value{
A raw_spct object containing the adjusted values, still as uncalibrated
  counts. The object is tagged with the with attribute "linearized" set to
  the function used for linearization.
}
\description{
Uses a function stored as an attribute of \code{x}, by default retrieved
from the instrument's firmware at the time of data acquisition or possibly
a replacement set by the user.
}
\note{
In contrast to other classes defined in package 'photobiology', class
  "raw_spct" can have more than one column of raw counts in cases where the
  intention is to merge these values as part of the processing at the time
  the calibration is applied. In other words these columns are not separate
  observations but instrumental replicates or bracketing readings part of the
  same "logical" or "effective" observation. The contents of any column whose
  name starts with "counts" will have the intrument response linearization
  function applied.
}
\seealso{
Other spectral data-processing functions:
\code{\link[=MAYP112785_tail_correction]{MAYP112785_tail_correction()}},
\code{\link[=MAYP11278_tail_correction]{MAYP11278_tail_correction()}},
\code{\link[=check_sn_match]{check_sn_match()}},
\code{\link[=merge_raw_mspct]{merge_raw_mspct()}},
\code{\link[=new_correction_method]{new_correction_method()}},
\code{\link[=ref_correction]{ref_correction()}},
\code{\link[=trim_counts]{trim_counts()}},
\code{\link[=uvb_corrections]{uvb_corrections()}}
}
\concept{spectral data-processing functions}
