% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{list_srs_interactive}
\alias{list_srs_interactive}
\title{Get list of connected instruments}
\usage{
list_srs_interactive(w)
}
\arguments{
\item{w}{handle to Omni Driver, used to test if an spectrometer is still
connected.}
}
\description{
Get list of spectrometers requesting user to connect one or abort if none
found.
}
\seealso{
Other interactive acquisition utility functions:
\code{\link[=choose_ch_interactive]{choose_ch_interactive()}},
\code{\link[=choose_sr_interactive]{choose_sr_interactive()}},
\code{\link[=f.trigger.message]{f.trigger.message()}},
\code{\link[=protocol_interactive]{protocol_interactive()}},
\code{\link[=set_attributes_interactive]{set_attributes_interactive()}},
\code{\link[=set_folder_interactive]{set_folder_interactive()}},
\code{\link[=set_seq_interactive]{set_seq_interactive()}},
\code{\link[=set_session_name_interactive]{set_session_name_interactive()}},
\code{\link[=set_user_name_interactive]{set_user_name_interactive()}},
\code{\link[=set_wlrange_interactive]{set_wlrange_interactive()}},
\code{\link[=tune_interactive]{tune_interactive()}}
}
\concept{interactive acquisition utility functions}
