% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-oo-header.R
\name{map_oofile_header_rows}
\alias{map_oofile_header_rows}
\title{Parse a file header to locate metadata items}
\usage{
map_oofile_header_rows(lines, header.end = NULL, grammar = oo.minimum.gr)
}
\arguments{
\item{lines}{character The data file read line by line.}

\item{header.end}{integer The index to the last line of the longest
header expected.}

\item{grammar}{data.frame With character variables "feature" and "pattern"
that will be used to locate the lines containing each type of metadata.}
}
\value{
A data frame with two variables, "feature" of class character, and
  "line.idx" of class integer.
}
\description{
Locate in which lines of the file header different metadata features are
located. Parsing is needed because the format used varies with among Ocean
Optics programs and instruments. Even the number of lines in the header, and
the position of the first line with spectral data vary.
}
\seealso{
Other functions for importing spectral data from files:
\code{\link[=collect_spct_files]{collect_spct_files()}},
\code{\link[=oofile_data_rows]{oofile_data_rows()}},
\code{\link[=plot_spct_file]{plot_spct_file()}},
\code{\link[=read_files2mspct]{read_files2mspct()}},
\code{\link[=read_oo_data]{read_oo_data()}},
\code{\link[=read_oo_ovdata]{read_oo_ovdata()}},
\code{\link[=read_oo_pidata]{read_oo_pidata()}},
\code{\link[=read_oo_ssdata]{read_oo_ssdata()}},
\code{\link[=set_oo_ssdata_descriptor]{set_oo_ssdata_descriptor()}},
\code{\link[=set_oo_ssdata_settings]{set_oo_ssdata_settings()}}
}
\concept{functions for importing spectral data from files}
