% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{maya_descriptor}
\alias{maya_descriptor}
\title{Instrument descriptor for a Maya2000 Pro  spectrometer}
\format{A list.}
\usage{
maya_descriptor
}
\description{
This a list with data needed to read the example raw data files
}
\keyword{datasets}

