% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maya_tail_correction}
\alias{maya_tail_correction}
\title{Function to compute the tail correction}
\usage{
maya_tail_correction(w_length, cts_second)
}
\arguments{
\item{w_length}{numeric vector of wavelengths in nm.}

\item{cts_second}{numeric vector of counts per second (stray-light corrected).}
}
\value{
tail returned as numeric vector within a list.

Tail correction is a reimplementation of the calculations developed by Lasse
Ylianttila (STUK, Finland), originally in Excel.
}
\description{
Function to compute the tail correction
}
\author{
Pedro J. Aphalo
}
\references{
\url{http://www.r4photobiology.info}
}

