% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{maya_ylianttila}
\alias{maya_ylianttila}
\title{Correction method data values}
\format{A list.}
\usage{
maya_ylianttila
}
\description{
This a list with data needed to process raw data in the example files
}
\keyword{datasets}

