% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP112785_calib_dates.df}
\alias{MAYP112785_calib_dates.df}
\title{Dates for which the different descriptors conatin valid calibrations.}
\format{A list.}
\usage{
MAYP112785_calib_dates.df
}
\description{
This a data frame with data needed to process raw data acquired on different
dates.
}
\keyword{datasets}

