% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP112785_ylianttila.mthd}
\alias{MAYP112785_ylianttila.mthd}
\title{Correction method definition}
\format{A list.}
\usage{
MAYP112785_ylianttila.mthd
}
\description{
This a list with data needed to process raw data in the example files. It
contains constants and functions defining the method developed by Lasse
Ylianttila.
}
\keyword{datasets}
