% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP11278_calib_dates.df}
\alias{MAYP11278_calib_dates.df}
\title{Dates for which specific calibrations are valid.}
\format{
A list.
}
\usage{
MAYP11278_calib_dates.df
}
\description{
Data frame with data used to match raw spectral data to different
calibrations based on the date data was acquired.
}
\seealso{
Other objects containing instrument-specific data: 
\code{\link{MAYP112785_calib_dates.df}},
\code{\link{MAYP112785_descriptors}},
\code{\link{MAYP112785_simple.mthd}},
\code{\link{MAYP112785_sun.mthd}},
\code{\link{MAYP112785_ylianttila.mthd}},
\code{\link{MAYP11278_descriptors}},
\code{\link{MAYP11278_simple.mthd}},
\code{\link{MAYP11278_sun.mthd}},
\code{\link{MAYP11278_ylianttila.mthd}},
\code{\link{MAYP114590_cal.spct}},
\code{\link{MAYP114590_calib_dates.df}},
\code{\link{MAYP114590_descriptors}},
\code{\link{MAYP114590_none.mthd}},
\code{\link{MAYP114590_simple.mthd}},
\code{\link{MAYP114590_sun.mthd}},
\code{\link{MAYP114590_ylianttila.mthd}}
}
\concept{objects containing instrument-specific data}
\keyword{datasets}
