% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP114590_none.mthd}
\alias{MAYP114590_none.mthd}
\title{Correction method definition}
\format{
A list.
}
\usage{
MAYP114590_none.mthd
}
\description{
List with data needed to process raw data in the example files. It
contains constants and functions defining a method implementing only
minimal corrections for stray light and no correction for the slit function.
}
\seealso{
Other objects containing instrument-specific data: 
\code{\link{MAYP112785_calib_dates.df}},
\code{\link{MAYP112785_descriptors}},
\code{\link{MAYP112785_simple.mthd}},
\code{\link{MAYP112785_sun.mthd}},
\code{\link{MAYP112785_ylianttila.mthd}},
\code{\link{MAYP11278_calib_dates.df}},
\code{\link{MAYP11278_descriptors}},
\code{\link{MAYP11278_simple.mthd}},
\code{\link{MAYP11278_sun.mthd}},
\code{\link{MAYP11278_ylianttila.mthd}},
\code{\link{MAYP114590_cal.spct}},
\code{\link{MAYP114590_calib_dates.df}},
\code{\link{MAYP114590_descriptors}},
\code{\link{MAYP114590_simple.mthd}},
\code{\link{MAYP114590_sun.mthd}},
\code{\link{MAYP114590_ylianttila.mthd}}
}
\concept{objects containing instrument-specific data}
\keyword{datasets}
