% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-irrad-interactive.R
\name{mspct_summary}
\alias{mspct_summary}
\title{Summarize spectral irradiance}
\usage{
mspct_summary(
  mspct,
  unit.out = getOption("photobiology.radiation.unit", default = "energy"),
  scale.factor = ifelse(unit.out == "photon", 1e+06, 1),
  type = "plant",
  digits = 3L
)
}
\arguments{
\item{unit.out}{character One of "photon" or "energy".}

\item{scale.factor}{numeric A multiplicative factor used to rescale data.}

\item{digits}{integer The number of significant digits in the output.}

\item{mscpt}{A source_mspct, or a source_spct object containing spectral
irradiance for one or more sources.}

\item{summary.type}{character One of "plant", "PAR" or "VIS".}
}
\value{
A dataframe or tibble with one row per spectrum and one column per
   summary quantity and attribute and a column with the names of the spectra.
}
\description{
Compute waveband irradiances and rattios between waveband irradiances of
interest to plants' and visual responses to light.
}
\details{
This function packages different functions from pacakge 'photobiology'
   and returns a typical set of summaries for different purposes.
}
