% Not generated by roxygen2: edit by hand
% Please edit this file in man/photobiology-files.Rd
\name{photobiology-files}
\alias{r4p-data-files}
\title{Example R data files}
\description{
Package 'ooacquire' includes example data files. R \code{.rda} files contain
spectral data acquired or imported using 'ooacquire'.}
\details{
When spectral data are acquired or imported with 'ooacquire' they are stored
in classes defined in package \code{\link[photobiology]{photobiology}} and
normally saved as R data files.
Examples of these files are used in the documentation and for testing functions
designed for data import and conversion of detector-raw-counts into calibrated
physical quantityy values. As it is possible to use many of the functions from
'ooacquire' off-line without a live connection to a spectrometer, testing when
binaries are built in R-Universe or locally during development is enabled in
many cases when no spectromter is attached. The files can be also useful when
learning to use 'ooacquire'.}
\examples{
# find where the R files are installed as part of the package
path <- system.file("extdata", package = "ooacquire")

# list the files in the folder
list.files(path, pattern = "*.Rda$")

# load R objects from the file
\dontrun{
load(paste(path, "sun001.spct.Rda", sep = "/"))
summary(sun001.spct)
summary(sun001.raw_mspct)}}
