% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{protocol_interactive}
\alias{protocol_interactive}
\title{Interactively select a measurement protocol}
\usage{
protocol_interactive(protocols, default = names(protocols)[[1]])
}
\arguments{
\item{protocols}{named list Measuring protocol defifinitions and names.}

\item{default}{character Name of the default protocol.}
}
\value{
The member vector corresponding to the protocol selected by
the user.
}
\description{
Choose a protocol by name from a list of protocols, allowing the user to
correct the selection if needed. Protocols are not hard-wired, but instead
defined by the list passed as argument to \code{protocols}.
}
\details{
A protocol is defined as a named list of character strings, and consist in
multiple acquisition of spectra contributing to the same logical measurement.
The name of the list member is the name of the protocol, while the members
of each character vector correspond to spectra to be acquired, after some
change in the measurement conditions. For example the list
\code{list(rsd = c("reference", "sample", "dark"), rs = c("reference", "sample"))}
defines two protocols.
}
\seealso{
Other interactive acquisition utility functions:
\code{\link[=choose_ch_interactive]{choose_ch_interactive()}},
\code{\link[=choose_sr_interactive]{choose_sr_interactive()}},
\code{\link[=f.trigger.message]{f.trigger.message()}},
\code{\link[=list_srs_interactive]{list_srs_interactive()}},
\code{\link[=set_attributes_interactive]{set_attributes_interactive()}},
\code{\link[=set_folder_interactive]{set_folder_interactive()}},
\code{\link[=set_seq_interactive]{set_seq_interactive()}},
\code{\link[=set_session_name_interactive]{set_session_name_interactive()}},
\code{\link[=set_user_name_interactive]{set_user_name_interactive()}},
\code{\link[=set_wlrange_interactive]{set_wlrange_interactive()}},
\code{\link[=tune_interactive]{tune_interactive()}}
}
\concept{interactive acquisition utility functions}
