% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-counts.R
\name{QC_dark}
\alias{QC_dark}
\title{Quality control of dark spectra}
\usage{
QC_dark(
  x,
  range = NULL,
  tol.margin = 0.5,
  max.hot = 10,
  max.cold = 10,
  spct.label = "Spectrum ",
  verbose = getOption("photobiology.verbose", default = TRUE)
)
}
\arguments{
\item{x}{cps_spct or raw_spct A spectrum measured in darkness.}

\item{range}{numeric A wavelength range [nm].}

\item{tol.margin}{numeric in [0..small integer] indicating the tolerance
margin as a fraction of the median counts.}

\item{max.hot, max.cold}{integer Maximum number of hot and cold pixels
accepted.}

\item{spct.label}{character A character string to use in message.}

\item{verbose}{logical If true a message is emitted in addition to returning
the outcome.}
}
\value{
A logical value.
}
\description{
Function used to check the number of pixels that deviate from the median.
}
\details{
The expectation is that in a spectrum measured in the dark, to be
  used as a reference, the variation among pixel counts is small. This also
  applies in some cases to ranges of pixels protected from radiation by a
  long pass or short pass filter. By default the whole spectrum is included
  in the QC check, but if needed an argument can be passed to \code{range}
  to select a smaller region.
}
