% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-oo-files.R
\name{read_files2mspct}
\alias{read_files2mspct}
\title{Read multiple files into raw_mspct object}
\usage{
read_files2mspct(files, read.f = ooacquire::read_oo_data, ...)
}
\arguments{
\item{files}{a named list of character strings of file names}

\item{read.f}{a function which expects in its first parameter a file name or
file path}

\item{...}{additional arguments passed by name to the function passed as
argument to \code{read.f}}
}
\description{
Read multiple files and return a collection of raw spectra as a raw_spct
object.
}
\note{
Depending of the function passed to \code{read.f} more or less complete
metadata information will be stored as attributes in the raw_spct objects.
}
\seealso{
Other functions for importing spectral data from files:
\code{\link[=collect_spct_files]{collect_spct_files()}},
\code{\link[=map_oofile_header_rows]{map_oofile_header_rows()}},
\code{\link[=oofile_data_rows]{oofile_data_rows()}},
\code{\link[=plot_spct_file]{plot_spct_file()}},
\code{\link[=read_oo_data]{read_oo_data()}},
\code{\link[=read_oo_ovdata]{read_oo_ovdata()}},
\code{\link[=read_oo_pidata]{read_oo_pidata()}},
\code{\link[=read_oo_ssdata]{read_oo_ssdata()}},
\code{\link[=set_oo_ssdata_descriptor]{set_oo_ssdata_descriptor()}},
\code{\link[=set_oo_ssdata_settings]{set_oo_ssdata_settings()}}
}
\concept{functions for importing spectral data from files}
