% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{read_numbers}
\alias{read_numbers}
\title{Read numbers from the user}
\usage{
read_numbers(prompt, n.max = 1L, pattern = "[,;]")
}
\arguments{
\item{prompt}{character Prompt displayed at the console.}

\item{n.max}{integer Maximum vector length to return.}

\item{pattern}{character Passed to \code{gsub()}. Characters matched are
substituted by a space.}
}
\description{
A character string is read from the command prompt. The string is split at
white space and converted into \code{numeric}. If conversion fails NAs are
returned. If the \code{numeric} vector is longer than \code{n.max} it is
truncated.
}
\keyword{internal}
