% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-calibration.R
\name{read_oo_caldata}
\alias{read_oo_caldata}
\title{Read OO irradiance calibration.}
\usage{
read_oo_caldata(file, time = NULL, geocode = NULL, label = NULL,
  descriptor = NULL, tz = NULL, locale = NULL,
  verbose = getOption("photobiology.verbose", default = FALSE))
}
\arguments{
\item{file}{character string}

\item{time}{a \code{POSIXct} object, but if \code{NULL} the date stored in
file is used, and if \code{NA} no date variable is added}

\item{geocode}{A data frame with columns \code{lon} and \code{lat}.}

\item{label}{character string, but if \code{NULL} the value of \code{file} is
used, and if \code{NA} the "what.measured" attribute is not set.}

\item{descriptor}{list A list describing the instrument used.}

\item{tz}{character Time zone is by default read from the file.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{verbose}{Logical indicating the level of warnings wanted.}
}
\value{
A generic_spct object.
}
\description{
Reads and parses the header of a calibration data file as supplied by
Ocean Optics. The whole header is stored as a comment. The time is
retrieved and decoded.
}
