% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-ooov-files.R
\name{read_oo_ovdata}
\alias{read_oo_ovdata}
\title{Read File Saved by Ocean Optics' OceanView.}
\usage{
read_oo_ovdata(
  file,
  time = NULL,
  geocode = NULL,
  label = NULL,
  descriptor = NULL,
  tz = NULL,
  locale = NULL,
  verbose = getOption("photobiology.verbose", default = FALSE)
)
}
\arguments{
\item{file}{character string}

\item{time}{a \code{POSIXct} object, but if \code{NULL} the date stored in
file is used, and if \code{NA} no date variable is added}

\item{geocode}{A data frame with columns \code{lon} and \code{lat}.}

\item{label}{character string, but if \code{NULL} the value of \code{file} is
used, and if \code{NA} the "what.measured" attribute is not set.}

\item{descriptor}{list A list describing the instrument used.}

\item{tz}{character Time zone is by default read from the file.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{verbose}{Logical indicating the level of warnings wanted.}
}
\value{
A raw_spct object.
}
\description{
Reads and parses the header of a processed data file as output by
OceanView to extract the whole header remark field. The time field is
retrieved and decoded.
}
\seealso{
Other functions for importing spectral data from files:
\code{\link[=collect_spct_files]{collect_spct_files()}},
\code{\link[=map_oofile_header_rows]{map_oofile_header_rows()}},
\code{\link[=oofile_data_rows]{oofile_data_rows()}},
\code{\link[=plot_spct_file]{plot_spct_file()}},
\code{\link[=read_files2mspct]{read_files2mspct()}},
\code{\link[=read_oo_data]{read_oo_data()}},
\code{\link[=read_oo_pidata]{read_oo_pidata()}},
\code{\link[=read_oo_ssdata]{read_oo_ssdata()}},
\code{\link[=set_oo_ssdata_descriptor]{set_oo_ssdata_descriptor()}},
\code{\link[=set_oo_ssdata_settings]{set_oo_ssdata_settings()}}
}
\concept{functions for importing spectral data from files}
