% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-oopi-file.r
\name{read_oo_pidata}
\alias{read_oo_pidata}
\title{Read File Saved by Ocean Optics' Raspberry Pi software.}
\usage{
read_oo_pidata(
  file,
  date = NULL,
  geocode = NULL,
  label = NULL,
  tz = NULL,
  locale = readr::default_locale(),
  descriptor = NULL,
  corr.sensor.nl = FALSE,
  spectrometer.name = "Unknown via Raspberry Pi",
  spectrometer.sn = "Unknown via Raspberry Pi",
  npixels = 2048
)
}
\arguments{
\item{file}{character string}

\item{date}{a \code{POSIXct} object, but if \code{NULL} file modification
date is used with a warning and if \code{NA} date is set to NA.}

\item{geocode}{A data frame with columns \code{lon} and \code{lat}.}

\item{label}{character string, but if \code{NULL} the value of \code{file} is
used, and if \code{NA} the "what.measured" attribute is not set.}

\item{tz}{character Time zone is not saved to the file.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{descriptor}{list as returned by function \code{get_oo_descriptor}.}

\item{corr.sensor.nl}{logical, indicating if spectral data is already
linearized. If TRUE the spectrum is marked as linearized, and
linearization skipped during processing.}

\item{spectrometer.name, spectrometer.sn}{character.}

\item{npixels}{integer Number of pixels in spectral data.}
}
\value{
A raw_spct object.
}
\description{
Reads and parses the header of a raw data file as output by the server
running on a Raspberry Pi board to extract the whole header remark field. The
time field is retrieved and decoded.
}
\note{
The header in these files has very little information, so the user
needs to supply the number of pixels in the array as well as the date-time.
The file contains a date in milliseconds but as the Raspberry Pi board
contains no real-time clock, it seems to default to number of milliseconds
since the Pi was switched on.
}
\references{
\url{http://www.r4photobiology.info}
}
\seealso{
Other functions for importing spectral data from files:
\code{\link[=collect_spct_files]{collect_spct_files()}},
\code{\link[=map_oofile_header_rows]{map_oofile_header_rows()}},
\code{\link[=oofile_data_rows]{oofile_data_rows()}},
\code{\link[=plot_spct_file]{plot_spct_file()}},
\code{\link[=read_files2mspct]{read_files2mspct()}},
\code{\link[=read_oo_data]{read_oo_data()}},
\code{\link[=read_oo_ovdata]{read_oo_ovdata()}},
\code{\link[=read_oo_ssdata]{read_oo_ssdata()}},
\code{\link[=set_oo_ssdata_descriptor]{set_oo_ssdata_descriptor()}},
\code{\link[=set_oo_ssdata_settings]{set_oo_ssdata_settings()}}
}
\concept{functions for importing spectral data from files}
