% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{read_period}
\alias{read_period}
\title{Read period from the user}
\usage{
read_period(prompt, n.max = 1L, pattern = "[,;]")
}
\arguments{
\item{prompt}{character Prompt displayed at the console.}

\item{n.max}{integer Maximum vector length to return.}

\item{pattern}{character Passed to \code{gsub()}. Characters matched are
substituted by a space.}
}
\value{
A \code{period} vector of lengths of time.
}
\description{
A character string is read from the command prompt. The string is split at
white space and converted into \code{period}. If conversion fails NAs are
returned. If the \code{period} vector is longer than \code{n.max} it is
truncated.
}
\keyword{internal}
