% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-raw-data.R
\docType{data}
\name{red_filter.raw_mspct}
\alias{red_filter.raw_mspct}
\title{Raw counts data for a filter measurement}
\format{
A raw_mspct
}
\usage{
red_filter.raw_mspct
}
\description{
A red long-pass filter from Heliopan (695 nm) measured using a household
incandescent lamp as light source. The spectrometer used was an Ocean Optics
Maya2000 Pro.
}
\seealso{
Other objects containing example raw-counts data:
\code{\link{blue_filter.raw_mspct}},
\code{\link{halogen.raw_mspct}},
\code{\link{sun001.raw_mspct}},
\code{\link{white_grow_LED.raw_mspct}},
\code{\link{xenon_flash.raw_mspct}}
}
\concept{objects containing example raw-counts data}
\keyword{datasets}
