% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref-correction.R
\name{ref_correction}
\alias{ref_correction}
\alias{ref_correction.default}
\alias{ref_correction.numeric}
\alias{ref_correction.raw_spct}
\alias{ref_correction.cps_spct}
\alias{ref_correction.cps_mspct}
\title{Apply a correction to spectral data.}
\usage{
ref_correction(x, y, .oper, ...)

\method{ref_correction}{default}(x, y, .oper, ...)

\method{ref_correction}{numeric}(x, y, .oper = `-`, ...)

\method{ref_correction}{raw_spct}(x, y, .oper = `-`, ...)

\method{ref_correction}{cps_spct}(x, y, .oper = `-`, ...)

\method{ref_correction}{cps_mspct}(x, y = x, .oper = `-`, ref_name = "dark", ...)
}
\arguments{
\item{x, y}{R objects}

\item{.oper}{a function with its first two formal parameters accepting
numerical vectors of equal length (e.g. a binary numerical operator).}

\item{...}{not used in current version}

\item{ref_name}{character Name of variable to substract from all other
columns.}
}
\value{
a numeric vector of the same length as x

an object of class "cps_spct"

an object of class "cps_spct"

an object of class "cps_mspct"
}
\description{
These functions simply subtract (by default) detector counts data (raw or cps)
of one spectrum by the corresponding columns in another spectrum, or apply
a user supplied operator or function to them.
}
\section{Methods (by class)}{
\itemize{
\item \code{ref_correction(default)}: Default method

\item \code{ref_correction(numeric)}: Numeric method

\item \code{ref_correction(raw_spct)}: Method for spectral data expressed as raw
instrument counts.

\item \code{ref_correction(cps_spct)}: Method for spectral data expressed as
counts per second.

\item \code{ref_correction(cps_mspct)}: Method for collections of spectral data objects
containing data expressed as counts per second.

}}
\note{
In the case of objects of class "raw_spct" the columns with names
starting with "counts" are processed. All other columns are left unchanded.

If \code{x} and \code{y} are both cps_mspct objects,
  \code{y[[ref_name]]} will be used as reference, otherwise \code{y} itself
  should be a cps_spct and will be used as is. In all cases variables in
  ref.name will be skipped in \code{x}.
}
\seealso{
Other spectral data-processing functions:
\code{\link[=MAYP112785_tail_correction]{MAYP112785_tail_correction()}},
\code{\link[=MAYP11278_tail_correction]{MAYP11278_tail_correction()}},
\code{\link[=check_sn_match]{check_sn_match()}},
\code{\link[=linearize_counts]{linearize_counts()}},
\code{\link[=merge_raw_mspct]{merge_raw_mspct()}},
\code{\link[=new_correction_method]{new_correction_method()}},
\code{\link[=trim_counts]{trim_counts()}},
\code{\link[=uvb_corrections]{uvb_corrections()}}
}
\concept{spectral data-processing functions}
