% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s-irrad-corrected-updt.R
\name{s_irrad_corrected_updt}
\alias{s_irrad_corrected_updt}
\title{Compute spectral irradiance update}
\usage{
s_irrad_corrected_updt(
  x,
  y = NULL,
  which = NULL,
  which.not = NULL,
  verbose = getOption("photobiology.verbose", default = FALSE)
)
}
\arguments{
\item{x}{A named list of one to three vectors of file names, with names
"light", "filter", and "dark". Or a raw_mspt object, or a raw_spct object.}

\item{y}{source_spct Object containing spectral irradiance earlier
computed from the same \code{raw_mspct} object as passed to \code{x}.}

\item{which, which.not}{character Names of attributes to copy or
not to copy.}

\item{verbose}{logical Enable informative messages.}
}
\value{
A \code{source_spct} or a \code{cps_spct} object.
}
\description{
Compute spectral irradiance using the instrument descriptor
for the same spectrometer and date of measurement using the
calibration data in the current version of 'ooacquire' rather
than that available at the time when raw data were acquired.
}
\details{
When raw counts data are acquired with
  \code{acq_irrad_interactive()} or imported from text files from other
  software the calibration information is stored together with the data. When
  computing spectral irradiance with function \code{s_irrad_corrected()} at
  any later time, this stored calibration data are the ones used. This
  approach ensures reproducibility as calibrations can be, and have been,
  retrospectively modified when 'ooacquire' is updated. In most cases the
  updates have only added information about "hot" and "dead" pixels. However,
  in one case the range of the calibration data included was incomplete, and
  later expanded. These changes can be important in some cases, making it
  necessary or preferable to use the updated calibrations when computing
  spectral irradiance.

  In \code{s_irrad_corrected_updt()} the \code{instr.desc} attribute in \code{x} is matched
  based on spectrometer serial number, date and entrance optics
  to the current, and possibly updated, version of the same calibration
  data in the current version of 'ooacquire'. The instrument descriptor
  is replaced in a local copy of \code{x} and this local copy used to
  computed spectral irradiance with \code{s_irrad_corrected()}, without
  altering \code{x}.

  If an argument is passed to parameter \code{y} and the value of the
  \code{when.measured} attribute and spectrometer serial number match between
  \code{x} and \code{y}, attributes, possibky added or modified after
  \code{y} was computed, are copied to the returned \code{source_spct} object
  using method \code{\link[photobiology]{copy_attributes}()} passing
  \code{which} and \code{which.not} in the call. This creates, as close as
  possible an updated version of \code{y}. Attributes \code{"instr.desc"},
  \code{"instr.settings"}, \code{"when.measured"} and several others set by
  \code{s_irrad_corrected()} are never updates in the returned value to
  ensure they remain valid. Other attributes including
  \code{"what.measured"}, \code{"where.measured"}, \code{"how.measured"},
  \code{"comment"}, are copied by default. If names are passed in an argument
  to \code{which} only these attributes are copied, with user-set attributes
  allowed. However, exclusion of attributes named in the union of the
  argument passed to \code{which.not} and those not excluded by default takes
  precedence.
}
\examples{

# only metadata from x and the new descriptor
s_irrad_corrected_updt(x = white_grow_LED.raw_mspct)

}
\seealso{
copy_attributes

Other functions for conversion of raw-counts data:
\code{\link[=raw2corr_cps]{raw2corr_cps()}},
\code{\link[=raw2cps]{raw2cps()}},
\code{\link[=s_fraction_corrected]{s_fraction_corrected()}},
\code{\link[=s_irrad_corrected]{s_irrad_corrected()}}
}
\concept{functions for conversion of raw-counts data}
