% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s-irrad-refresh.R
\name{s_irrad_refresh}
\alias{s_irrad_refresh}
\title{Recompute spectral irradiance}
\usage{
s_irrad_refresh(x, old.spct = NULL)
}
\arguments{
\item{x}{A named list of one to three vectors of file names, with names
"light", "filter", and "dark". Or a raw_mspt object, or a raw_spct object.}

\item{old.spct}{source_spct Object from which to copy metadata.}
}
\description{
Recompute spectral irradiance using the instrument descriptor
for the same spectrometer and date of measurement using the
calibration data in the current version of 'ooacquire' rather
than that available at the time tha raw data were acquired.
}
\examples{

s_irrad_refresh(white_LED.raw_mspct, white_led.source_spct)


}
