% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_entrance_optics}
\alias{set_descriptor_entrance_optics}
\title{Add or replace entrance optics data to descriptor}
\usage{
set_descriptor_entrance_optics(
  descriptor,
  make = NA_character_,
  model = NA_character_,
  geometry = NA_character_,
  serial.number = NA_character_,
  area = NA_real_
)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{make, model, geometry, serial.number}{character.}

\item{area}{numeric (\eqn{m^2}).}
}
\value{
a copy of the argument passed for \code{oo_descriptor} with the
\code{entrance.optics} field of the calibration data adde or replace
by the new list.
}
\description{
Replace or add a description of the fibre and diffuser or other entrance
optics to the instrument descriptor.
}
\seealso{
Other Functions and methods related to instrument descriptors:
\code{\link[=get_oo_descriptor]{get_oo_descriptor()}},
\code{\link[=set_descriptor_bad_pixs]{set_descriptor_bad_pixs()}},
\code{\link[=set_descriptor_integ_time]{set_descriptor_integ_time()}},
\code{\link[=set_descriptor_irrad_mult]{set_descriptor_irrad_mult()}},
\code{\link[=set_descriptor_nl]{set_descriptor_nl()}},
\code{\link[=set_descriptor_wl]{set_descriptor_wl()}}
}
\concept{Functions and methods related to instrument descriptors}
