% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_slit_fun}
\alias{set_descriptor_slit_fun}
\title{Add function for slit correction}
\usage{
set_descriptor_slit_fun(descriptor, inv.slit.fun)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{inv.slit.fun}{function with two first formal parameters taking numeric
vectors of wavelengths and counts that aoolies a correction suitable
for the instrument.}
}
\value{
A copy of the argument passed for \code{oo_descriptor} with the
irrad.mult field of the calibration data replaced by the new values.
}
\description{
Adds a function to the instrument descriptor that can be used to
correct counts per second data to remove the effect of the tails of
the slit function of the instrument.
}

