% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_wl}
\alias{set_descriptor_wl}
\title{Replace wavelength values in an instrument description}
\usage{
set_descriptor_wl(descriptor, wl)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{wl}{numeric vector of sorted wavelengths values corresponding to each
pixel in the instrument array.}
}
\value{
a copy of the argument passed for \code{oo_descriptor} with the
wavelengths field of the calibration data replaced by the new values.
}
\description{
Replace wavelength values in an instrument descriptor for an Ocean Optics
spectrometer with new values. (e.g. when wavelength calibration is not stored
in firmware).
}
\seealso{
Other Functions and methods related to instrument descriptors:
\code{\link[=get_oo_descriptor]{get_oo_descriptor()}},
\code{\link[=set_descriptor_bad_pixs]{set_descriptor_bad_pixs()}},
\code{\link[=set_descriptor_entrance_optics]{set_descriptor_entrance_optics()}},
\code{\link[=set_descriptor_integ_time]{set_descriptor_integ_time()}},
\code{\link[=set_descriptor_irrad_mult]{set_descriptor_irrad_mult()}},
\code{\link[=set_descriptor_nl]{set_descriptor_nl()}}
}
\concept{Functions and methods related to instrument descriptors}
