% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-settings.R
\name{set_integ_time}
\alias{set_integ_time}
\title{Set integration time}
\usage{
set_integ_time(acq.settings, integ.time, single.scan = FALSE, verbose = TRUE)
}
\arguments{
\item{acq.settings}{list as returned by function \code{acq_settings()}}
}
\value{
a copy of the argument passed for \code{acq.settings} with the
\code{integ.time} field of the settings data replaced by \code{integ.time}.
}
\description{
Set the integration time in the instrument settings data structure.
}
\seealso{
Other acquisition-settings related functions:
\code{\link[=acq_settings]{acq_settings()}},
\code{\link[=set_linearized]{set_linearized()}},
\code{\link[=set_num_exposures]{set_num_exposures()}}
}
\concept{acquisition-settings related functions}
\keyword{internal}
