% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-settings.R
\name{set_linearized}
\alias{set_linearized}
\title{Set linearized attribute}
\usage{
set_linearized(acq.settings)
}
\arguments{
\item{acq.settings}{list as returned by function \code{acq_settings()}}
}
\value{
a copy of the argument passed for \code{acq.settings} with the
linearized field of the settings data replaced by TRUE.
}
\description{
Tag the settings as linearized
}
\seealso{
Other acquisition-settings related functions:
\code{\link[=acq_settings]{acq_settings()}},
\code{\link[=set_integ_time]{set_integ_time()}},
\code{\link[=set_num_exposures]{set_num_exposures()}}
}
\concept{acquisition-settings related functions}
\keyword{internal}
