% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-oo-header.R
\name{set_oo_ssdata_settings}
\alias{set_oo_ssdata_settings}
\title{Set the values of instrument settings from file header}
\usage{
set_oo_ssdata_settings(x, file.header = comment(x), overwrite = TRUE)
}
\arguments{
\item{x}{generic_spct, although with defaults only raw_spct.}

\item{file.header}{character string The header of the file output by
SpectraSuite.}

\item{overwrite}{logical A flag indicating if an existing valid instrument
descriptor in \code{x} should be overwritten.}
}
\value{
A copy of \code{x} with updated attributes.
}
\description{
Parse the header of the file returned by SpectraSuite for corrections,
smoothing and acquisition parameters used. These values are used to set
the "inst.settings" attribute of \code{x}.
}
\seealso{
Other functions for importing spectral data from files:
\code{\link[=collect_spct_files]{collect_spct_files()}},
\code{\link[=map_oofile_header_rows]{map_oofile_header_rows()}},
\code{\link[=oofile_data_rows]{oofile_data_rows()}},
\code{\link[=plot_spct_file]{plot_spct_file()}},
\code{\link[=read_files2mspct]{read_files2mspct()}},
\code{\link[=read_oo_data]{read_oo_data()}},
\code{\link[=read_oo_ovdata]{read_oo_ovdata()}},
\code{\link[=read_oo_pidata]{read_oo_pidata()}},
\code{\link[=read_oo_ssdata]{read_oo_ssdata()}},
\code{\link[=set_oo_ssdata_descriptor]{set_oo_ssdata_descriptor()}}
}
\concept{functions for importing spectral data from files}
