% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{set_wlrange_interactive}
\alias{set_wlrange_interactive}
\title{Interactively set plot wavelength range}
\usage{
set_wlrange_interactive(wl.range.boundaries)
}
\arguments{
\item{wl.range.boundaries}{numeric vector of length two. Maximum allowed
wavelength range in nanometres (nm).}
}
\value{
a named list of character vectors.
}
\description{
Allow the user to set the range interactively by setting R option
\code{ggspectra.wlrange}.
}
\seealso{
Other interactive acquisition utility functions:
\code{\link[=choose_ch_interactive]{choose_ch_interactive()}},
\code{\link[=choose_sr_interactive]{choose_sr_interactive()}},
\code{\link[=f.trigger.message]{f.trigger.message()}},
\code{\link[=list_srs_interactive]{list_srs_interactive()}},
\code{\link[=protocol_interactive]{protocol_interactive()}},
\code{\link[=set_attributes_interactive]{set_attributes_interactive()}},
\code{\link[=set_folder_interactive]{set_folder_interactive()}},
\code{\link[=set_seq_interactive]{set_seq_interactive()}},
\code{\link[=set_session_name_interactive]{set_session_name_interactive()}},
\code{\link[=set_user_name_interactive]{set_user_name_interactive()}},
\code{\link[=tune_interactive]{tune_interactive()}}
}
\concept{interactive acquisition utility functions}
\keyword{internal}
