% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-counts.R
\name{skip_bad_pixs}
\alias{skip_bad_pixs}
\title{Replace bad pixels}
\usage{
skip_bad_pixs(x)
}
\arguments{
\item{x}{raw_spct object}
}
\value{
a copy of x with values replaced as needed in all counts columns
  present.
}
\description{
Replace bad pixels with the average of the counts from the two neighbouring
pixels.
}
