% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrometer-defaults.R
\name{spectrometer_defaults}
\alias{spectrometer_defaults}
\title{Fetch spectrometer defaults}
\usage{
spectrometer_defaults(
  serial_no,
  entrance.optics = "cosine",
  data.acq.date = lubridate::now(tzone = "UTC"),
  descriptors = NA,
  correction.method = NA
)
}
\arguments{
\item{serial_no}{character The serial number of an Ocean Optics spectrometer
for which data is available in 'ooacquire'.}

\item{entrance.optics}{character, name or geometry of diffuser, needed only
if there is more than one for the same instrument.}

\item{data.acq.date}{Any object that anytime::anydate() will decode as a date
or convert to a date. Used to select a descriptor containing calibration
data valid for a given day.}

\item{descriptors}{list A list of instrument descriptors containing
calibration data.}

\item{correction.method}{list The method to use when applying the calibration}
}
\description{
Find default instrument descriptor, correction method, and default
measuring protocol based on spectrometer serial number and entrance
optics.
}
\details{
When any of \code{protocols}, \code{descriptor} or
\code{correction.method} is not available (\code{NA}), the
missing value is replaced by a default based on the arguments
passed to \code{serial_no}, \code{entrance.optics}, and
\code{protocols}.
}
\examples{

str(spectrometer_defaults(serial_no = "MAYP114590"))

}
