% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-fraction-interactive.R
\name{Tfr_summary_table}
\alias{Tfr_summary_table}
\title{Summarize spectral transmittance}
\usage{
Tfr_summary_table(
  mspct,
  scale.factor = 1,
  attr2tb = "when.measured",
  summary.type = "VIS",
  digits = 3L
)
}
\arguments{
\item{mspct}{A filter_mspct, or a filter_spct object containing spectral
transmittance for one or more sources.}

\item{scale.factor}{numeric A multiplicative factor used to rescale data.}

\item{attr2tb}{character Vector with one or more of "when.measured",
"what.measured", "where.measured", "how.measured" and "comment".}

\item{summary.type}{character One of "plant", "PAR" or "VIS".}

\item{digits}{integer The number of significant digits in the output.}
}
\value{
A tibble with one row per spectrum and one column per
   summary quantity and attribute and a column with the names of the spectra.
}
\description{
Compute transmittance by waveband of interest to plants' and human visual
responses to light.
}
\details{
This function calls different functions from package 'photobiology'
   and returns a typical set of summaries.
}
\examples{

Tfr_summary_table(yellow_gel.spct)
Tfr_summary_table(yellow_gel.spct, attr2tb = c("what.measured", "where.measured"))
Tfr_summary_table(yellow_gel.spct, summary.type = "plant")
Tfr_summary_table(yellow_gel.spct, summary.type = "PAR")
Tfr_summary_table(yellow_gel.spct, summary.type = "VIS")

}
\seealso{
See the documentation for functions
  \code{\link[photobiology]{transmittance}},
  \code{\link[photobiology]{add_attr2tb}}
  and \code{\link[base]{signif}} which are called to build
  the summary table.
}
