% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-counts.R
\name{trim_counts}
\alias{trim_counts}
\title{Replace out-of-range instrument counts}
\usage{
trim_counts(x, range = c(NA, getInstrDesc(x)[["max.counts"]] - 1), fill = NA)
}
\arguments{
\item{x}{raw_spct object}

\item{range}{integer vector of length two}

\item{fill}{an integer value (including NA) to be used to replace the values
that are outside range.}
}
\value{
a copy of x with values replaced as needed in all counts columns
  present.
}
\description{
Trim out-of-range counts values in data stored in a "raw_spct" object. The
values are replaced with the supplied \code{fill} value.
}
\seealso{
Other spectral data-processing functions:
\code{\link[=MAYP112785_tail_correction]{MAYP112785_tail_correction()}},
\code{\link[=MAYP11278_tail_correction]{MAYP11278_tail_correction()}},
\code{\link[=check_sn_match]{check_sn_match()}},
\code{\link[=linearize_counts]{linearize_counts()}},
\code{\link[=merge_raw_mspct]{merge_raw_mspct()}},
\code{\link[=new_correction_method]{new_correction_method()}},
\code{\link[=ref_correction]{ref_correction()}},
\code{\link[=uvb_corrections]{uvb_corrections()}}
}
\concept{spectral data-processing functions}
