% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.spc.obj.R
\name{tune_acq_settings}
\alias{retune_acq_settings}
\alias{tune_acq_settings}
\title{Tune settings for measurement}
\usage{
tune_acq_settings(oo_descriptor, start.integ.time = 0.05,
  min.integ.time = -Inf, max.integ.time = Inf, min.tot.time = 2,
  HDR.mult = c(short = 1, long = 10), NR.flag = c(short = TRUE, long =
  TRUE), pix.selector = TRUE, verbose = TRUE)

retune_acq_settings(oo_descriptor, x, verbose = TRUE)
}
\arguments{
\item{oo_descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{start.integ.time}{numeric vaue in seconds}

\item{min.integ.time}{numeric vaue in seconds}

\item{max.integ.time}{numeric vaue in seconds}

\item{min.tot.time}{numeric vaue in seconds}

\item{HDR.mult}{a numeric vector with integ.time multipliers to be used for
"bracketing".}

\item{NR.flag}{a logical vector indicating whether a filter reading will be
used.}

\item{pix.selector}{a logical or numeric vector used as subscript to select
pixels}

\item{verbose}{a logical to enable or disable warnings}

\item{x}{list as returned by a previous call to \code{tune_acq_settings()} or
\code{retune_acq_settings()}.}
}
\value{
a list.
}
\description{
Find parameters for spectral measurements with a given measurements
protocol.
}
\details{
\code{tune_acq_settings()} is used usually for the fist measurement
  in a series as it starts the tuning of the acquisition parameters from an
  integration time suplied by the user or an arbitrary default.
  \code{retune_acq_settings()} instead does a similar tunning starting from a
  previously adjusted set of values. This is just a convenience function to
  simplify data acquisition code and speed-up data acquisition by refining a
  previously set value.
}
\note{
Ocean Optics spectrometers can be queried for the maximum and minimum
  supported integration times. This function modifies the user supplied
  values if outside these bounds. The defaults of -Inf and Inf force the use
  of the whole valid range of integration times for the connected intrument.
  \code{pixel.selector} can be used for two different purposes: to ignore
  bad pixels and to restrict integration-time tuning to the response from a
  range of pixels.
}
\keyword{manip}
\keyword{misc}

