% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{tune_interactive}
\alias{tune_interactive}
\title{Interactively adjust the integration time settings}
\usage{
tune_interactive(
  descriptor,
  acq.settings,
  start.int.time = 0.1,
  interface.mode = "auto"
)
}
\arguments{
\item{descriptor}{list Descriptor of the instrument, including wrapper to
Java object used to access the instrument.}

\item{acq.settings}{list containing starting values for instrument settings.}

\item{start.int.time}{numeric Integration time to use as starting guess when
tuning the settings.}

\item{interface.mode}{character One of "simple", "auto", or "manual".}
}
\description{
Adjust integration time settings, allowing the user to repeat the tuning,
and to change some of the parameters used for tuning such as total
compound integration time and integration time bracketing.
}
\details{
This function implements three different user interfaces: 1) "simple" is an
interface suitable for the most usual measurements using automatic tuning
of integration time and hides some of the
less frequently used options, 2) "auto" gives access to all available
options offering maximum flexibility when using automatic tuning of
integration time, and 3) "manual" supports use of fixed integration times
directly entered by the user.

Tuning of the integration time takes into account the range of times
supported by the connected instrument, read from the instrument descriptor.
The algorithm also makes use of the linearisation function when extrapolating
to guess the integration time needed. Initial (default) values are read from
\code{acq.settings} while \code{start.int.time} provides a default starting
value for integration time for tuning when the user chooses not to use the
value stored in \code{acq.settings}.
}
\seealso{
Other interactive acquisition utility functions:
\code{\link[=choose_ch_interactive]{choose_ch_interactive()}},
\code{\link[=choose_sr_interactive]{choose_sr_interactive()}},
\code{\link[=f.trigger.message]{f.trigger.message()}},
\code{\link[=list_srs_interactive]{list_srs_interactive()}},
\code{\link[=protocol_interactive]{protocol_interactive()}},
\code{\link[=set_attributes_interactive]{set_attributes_interactive()}},
\code{\link[=set_folder_interactive]{set_folder_interactive()}},
\code{\link[=set_seq_interactive]{set_seq_interactive()}},
\code{\link[=set_session_name_interactive]{set_session_name_interactive()}},
\code{\link[=set_user_name_interactive]{set_user_name_interactive()}},
\code{\link[=set_wlrange_interactive]{set_wlrange_interactive()}}
}
\concept{interactive acquisition utility functions}
