% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-instr-desc.R
\name{update_instr_desc}
\alias{update_instr_desc}
\title{Update whole instrument descriptor}
\usage{
update_instr_desc(x, which.descriptor.args = list())
}
\arguments{
\item{x}{raw_spct or raw_mspct object with attribute \code{instr.desc} set.}

\item{which.descriptor.args}{named list Of arguments to use to fetch the
replacement descriptor.}
}
\value{
a copy of \code{x} with an updated \code{instr.desc} attribute
  embedded.
}
\description{
Update the instrument descriptor embedded in \code{raw_spct} and
\code{raw_mspct} objects.
}
\details{
Recomputing of physical quantities from raw counts normally reuses the
embedded calibration data in \code{\link[photobiology]{raw_spct}} objects.
Function \code{update_inst_desc()} makes it possible to update the embedded
calibration before recomputing derived quantities.

Spectral objects, including those with raw counts data contain an
instrument descriptor when measured or imported using functions from
'ooacquire'. The descriptor can include calibration data as well
as details of the instrument used. If a calibration has been retroactively
modified after data acquisition, the descriptor can be refreshed in
existing \code{raw_spct} or \code{raw_mspct} objects. By default the
the version of the new descriptor is that in the loaded version of
'ooacquire', corresponding to the same instrument, data acquisition date
and entrance optics.

If for some reason, an instrument descriptor for the wrong spectrometer,
the wrong entrance optics or for a wrong data acquisition date, a non-matching
replacement can be used by passing suitable arguments through parameter
\code{which.descriptor.arg}. See \code{\link{which_descriptor}()} for the
details.
}
\note{
The replacement of the instrument descriptor is not supported for
  spectral data not expressed as raw counts, because this would create an
  inconsistency between the computed data and its metadata. To achieve such a
  change data in physical units have to be recomputed from raw instrument
  counts after updating the object containing the raw detector counts.
}
\section{Warning!}{
 This function is intended only to repair objects that
have corrupted information. Typical cases are acquisition using a computer
with wrong date settings, unrecognized instrument serial number, or
measurements done before the valid calibration was added to 'ooacquire'.

The special case when the only change to the calibration is adding newly
identified bad pixels in the sensor array, function
\code{\link{update_bad_pixs}} provides a "light weight" alternative.
}

