% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-raw-data.R
\docType{data}
\name{warm_white_LED.raw_mspct}
\alias{warm_white_LED.raw_mspct}
\title{Raw counts data for a lamp measurement.}
\format{A raw_mscpt}
\usage{
warm_white_LED.raw_mspct
}
\description{
A household Osram LED STAR CLASSIC A 60 lamp measured at short distance. The
spectrometer used was an Ocean Optics Maya2000 Pro.
}
\keyword{datasets}
