% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb-corrections.R
\name{which_descriptor}
\alias{which_descriptor}
\title{Select which instrument descriptor to use}
\usage{
which_descriptor(date = lubridate::today(),
  descriptors = ooacquire::MAYP11278_descriptors, verbose = FALSE, ...)
}
\arguments{
\item{date}{Any object that lubridate::as_date() will decode as a date or
convert to a date. Used only to override the selection of the descriptor
containing calibration data.}

\item{descriptors}{A named list of descriptors of the characteristics of
the spectrometer including calibration data.}

\item{verbose}{Logical indicating the level of warnings wanted.}

\item{...}{Currently ignored.}
}
\description{
Select from a list of instrument descriptors which one to use based on
date of measurement.
}

