% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-raw-data.R
\docType{data}
\name{white_LED_2min.raw_mspct}
\alias{white_LED_2min.raw_mspct}
\title{Raw counts data for a lamp measurement.}
\format{A raw_mspct}
\usage{
white_LED_2min.raw_mspct
}
\description{
A household Osram LED STAR CLASSIC A 60 lamp measured at short distance. The
spectrometer used was an Ocean Optics Maya2000 Pro. Measured using three
level bracketing and a 120 seconds total time per measurement through scan
averaging.
}
\keyword{datasets}
