% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive.R
\name{acq_irrad_interactive}
\alias{acq_irrad_interactive}
\title{Acquire spectra interactively}
\usage{
acq_irrad_interactive(tot.time.range = c(1, 30), target.margin = 0.1,
  HDR.mult = c(short = 1, long = 10),
  method = ooacquire::MAYP11278_ylianttila.mthd,
  descriptors = ooacquire::MAYP11278_descriptors)
}
\arguments{
\item{tot.time.range}{numeric vector Range of total times for a measurement
in seconds.}

\item{target.margin}{numeric (0..1) when tuning integration time, how big a
head space to leave.}

\item{HDR.mult}{numeric the integration time for each bracketed integration
as a multiplier of the set or tuned integration time.}

\item{method}{list The method to use when applying the calibration}

\item{descriptors}{list A list of instrument descriptors containing
calibration data.}
}
\description{
Function providing a simple interactive front-end to the functions in the
package. Mostly as example code that can be modified for diferent
uses.
}
\note{
The integration time is set automatically so that the peak number of
  counts is close to 1 - \code{target.margin} of maximum range of the
  instrument detector. The minmum \code{tot.time} is achieved by increasing
  the number of scans.
}
\examples{

\dontrun{
# requires an Ocean Optics spectrometer to be connected via USB
acquire_interactive()
}

}

