% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-continuous-interactive.R
\name{acq_irrad_interactive}
\alias{acq_irrad_interactive}
\alias{acq_fraction_interactive}
\alias{acq_rfr_tfr_interactive}
\title{Acquire spectra interactively}
\usage{
acq_irrad_interactive(
  tot.time.range = c(5, 15),
  target.margin = 0.1,
  HDR.mult = c(short = 1, long = 10),
  protocols = NULL,
  correction.method = NA,
  descriptors = NA,
  stray.light.method = "none",
  seq.settings = NULL,
  area = NULL,
  diff.type = NULL,
  qty.out = "irrad",
  save.pdfs = TRUE,
  save.summaries = TRUE,
  save.collections = TRUE,
  interface.mode = "auto",
  folder.name = paste("acq", qty.out, lubridate::today(), sep = "-"),
  user.name = Sys.info()[["user"]],
  session.name = paste(user.name, strftime(lubridate::now(), "\%Y.\%b.\%d_\%H.\%M.\%S"),
    sep = "_")
)

acq_fraction_interactive(
  tot.time.range = c(5, 15),
  target.margin = 0.2,
  HDR.mult = c(short = 1, long = 10),
  protocols = NULL,
  correction.method = NA,
  descriptors = NA,
  ref.value = 1,
  qty.out = "Tfr",
  type = "total",
  stray.light.method = "simple",
  save.pdfs = TRUE
)

acq_rfr_tfr_interactive(
  tot.time.range = c(10, 15),
  target.margin = 0.1,
  HDR.mult = c(1, 10),
  descriptors = NA,
  ref.value = 1,
  save.pdfs = TRUE,
  qty.out = "Tfr"
)
}
\arguments{
\item{tot.time.range}{numeric vector Range of total times for a measurement
in seconds.}

\item{target.margin}{numeric (0..1) when tuning integration time, how big a
head space to leave.}

\item{HDR.mult}{numeric the integration time for each bracketed integration
as a multiplier of the set or tuned integration time.}

\item{protocols}{named list of character vectors, or a character vector with
names of at least one member of the default list of protocols.}

\item{correction.method}{list The method to use when applying the calibration}

\item{descriptors}{list A list of instrument descriptors containing
calibration data.}

\item{stray.light.method}{character Used only when the correction method is
created on-the-fly.}

\item{seq.settings}{named list with numeric members \code{"step.delay"} and
\code{"num.steps"}.}

\item{area}{numeric Passed to \code{o_calib2irrad_mult()}.}

\item{diff.type}{character Passed to \code{o_calib2irrad_mult()}.}

\item{qty.out}{character One of "Tfr" (spectral transmittance as a fraction
of one), "irrad" (spectral irardiance), "cps" (counts per second), or "raw"
(raw sensor counts).}

\item{save.pdfs, save.summaries, save.collections}{logical Whether to save
plots to PDFs files or not, and collection summaries to csv files or not,
enable collections user interface or not..}

\item{interface.mode}{character One of "auto", "simple", "manual", "series",
"auto-attr", "simple-attr", "manual-attr", and "series-attr".}

\item{folder.name, session.name, user.name}{character Default name of the
folder used for output, and session and user names.}

\item{ref.value}{numeric or filter_spct/reflector_spct object.}

\item{type}{character Type of transmittance or reflectance measured.}
}
\value{
These functions return the acquired spectra through "side effects" as
  each spectrum is saved, both as raw counts data and optionally as spectral
  irradiance or counts-per-second  spectral data in an \code{.rda} file as
  objects of the classes defined in package 'photobiology'. Optionally, the
  plot for each spectrum is saved as a \code{.pdf} file, as well as summaries
  to a CSV file and joint plots to a \code{.pdf} file for each collection
  created. The value returned by the function is that from closing the
  connection to the spectrometer.
}
\description{
Functions providing a simple interactive front-end to the functions in the
package, also working as example code that can be modified for different
uses.
}
\details{
These functions can be useful for commonly done measurements but
  they also play the role of examples that users can modify according to
  their needs. They are all composed in a modular way from functions that can
  be reshuffled and combined with other functions to define new variations
  better suited to users' needs and tastes. Functions
  \code{acq_irrad_interactive()} and \code{acq_fraction_interactive} provide
  support the measurement of spectral irradiance of continuous light sources,
  and transmittance, reflectance and absorbance using continuous light
  sources, respectively.

  The default behaviour of the functions can be changed by passing different
  arguments through parameters, but for special use cases it could be best
  for users to define case-specific data acquisition functions from the same
  building blocks.

  Using these functions only requires an Ocean Optics spectrometer to be
  connected. The connection to the spectrometer and selection of channel,
  when relevant, is done from within these functions.

  The irradiance calibration will be retrieved from the spectrometer memory
  as a last resource if not supplied in any other way. Given that the factors
  are stored in a format that ignores the entrance optics, either the
  efefctive diffuser are in xxx should be passed to parameter \code{area} or
  a character string with the type of the diffuser passed to
  \code{diff.type}. If no irardiance calibration is available, counts per
  second (cps) or raw counts are the only options available.
}
\note{
The integration time is set automatically so that the peak number of
  counts is close to 1 - \code{target.margin} times the maximum of the range
  of the instrument detector. The minimum \code{tot.time} is achieved by
  increasing the number of scans. The default protocols are usually suitable,
  if new protocols are passed, each character vector must contain strings
  "light", "filter" and "dark", or "sample", "reference", and "dark",
  depending on the function. Plots are produced with functions from package
  'ggspectra' and respect the default annotations set with function
  \code{set_annotations_default()}, default wavebands set with function
  \code{set_w.band_default()}, and irradiance quantities set with
  \code{photon_as_default()}, and \code{energy_as_default()}.

The different interface modes are suitable for different types of
  measurements. The field separator in the written CSV file is according to
  the current locale. If needed use \code{readr::set_locale()} to change it
  before calling this function.

Calibration data needs in most cases to be imported into R and
  parameters entered for the special correction algorithms into a correction
  method descriptor. The corrections are skipped if the needed information is
  missing. If no spectral irradiance calibration is available and attempt is
  made to retrieve it from the spectrometer, but given the format used by
  Ocean Optics/Ocean Insight, in this case the effective \code{area} of the
  cosine diffuser used (or the model name if from Ocean Optics) should be
  supplied by the user.

The calculations for reflectance and transmittance are very similar,
  so we provide a single function capable of handling both. For
  transmittance the reference is usually direct exposure to radiation but
  for reflectance a white reference patch is normally used. In some cases
  one may want to use a grey reference. We provide an argument that allows
  the user to supply a constant or a spectrum describing the properties of
  the reference. It is also important to distinguish between total and
  internal transmittance, and between total and specular reflectance.
  In both cases which of these is measured depends on the measuring protocol
  (condition used as reference, use of an integrating sphere versus use of a
  probe with a narrow angle of aperture, etc.) and consequently the correct
  value should be entered to ensure that data are correctly tagged and
  later computations valid.
}
\examples{
# please, see also the example scripts installed with the package

\dontrun{
# requires an Ocean Insight (former Ocean Optics) spectrometer to be
# connected via USB

acquire_interactive()
acquire_interactive(qty.out = "cps")

}

}
\seealso{
This function calls functions \code{\link{tune_interactive}},
  \code{\link{protocol_interactive}} and
  \code{\link{set_attributes_interactive}}. If irradiance calibration is
  retrieved from the instrument, functions \code{\link{get_oo_descriptor}}
  and \code{\link{oo_calib2irrad_mult}} are also called.

Other interactive acquisition functions: 
\code{\link{acq_fluence_interactive}()}
}
\concept{interactive acquisition functions}
