% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-irrad-interactive.R
\name{acq_irrad_interactive}
\alias{acq_irrad_interactive}
\title{Acquire spectral irradiance}
\usage{
acq_irrad_interactive(
  tot.time.range = c(5, 15),
  target.margin = 0.1,
  HDR.mult = c(short = 1, long = 10),
  protocols = NULL,
  correction.method = NA,
  descriptors = NA,
  stray.light.method = "none",
  seq.settings = NULL,
  area = NULL,
  diff.type = NULL,
  qty.out = "irrad",
  save.pdfs = TRUE,
  save.summaries = TRUE,
  save.collections = TRUE,
  interface.mode = "auto",
  folder.name = paste("acq", qty.out, lubridate::today(tzone = "UTC"), sep = "-"),
  user.name = Sys.info()[["user"]],
  session.name = paste(user.name, strftime(lubridate::now(tzone = "UTC"),
    "\%Y.\%b.\%d_\%H.\%M.\%S"), sep = "_")
)
}
\arguments{
\item{tot.time.range}{numeric vector Range of total times for a measurement
in seconds.}

\item{target.margin}{numeric (0..1) when tuning integration time, how big a
head space to leave.}

\item{HDR.mult}{numeric the integration time for each bracketed integration
as a multiplier of the set or tuned integration time.}

\item{protocols}{named list of character vectors, or a character vector with
names of at least one member of the default list of protocols.}

\item{correction.method}{list The method to use when applying the calibration}

\item{descriptors}{list A list of instrument descriptors containing
calibration data.}

\item{stray.light.method}{character Used only when the correction method is
created on-the-fly.}

\item{seq.settings}{named list with numeric members \code{start.boundary},
\code{initial.delay}, \code{"step.delay"} and \code{"num.steps"}.}

\item{area}{numeric Passed to \code{o_calib2irrad_mult()}.}

\item{diff.type}{character Passed to \code{o_calib2irrad_mult()}.}

\item{qty.out}{character One of "Tfr" (spectral transmittance as a fraction
of one), "irrad" (spectral irardiance), "cps" (counts per second), or "raw"
(raw sensor counts).}

\item{save.pdfs, save.summaries, save.collections}{logical Whether to save
plots to PDFs files or not, and collection summaries to csv files or not,
enable collections user interface or not..}

\item{interface.mode}{character One of "auto", "simple", "manual", "full",
"series", "auto-attr", "simple-attr", "manual-attr", "full-atr", and
"series-attr".}

\item{folder.name, session.name, user.name}{character Default name of the
folder used for output, and session and user names.}
}
\value{
These functions return the acquired spectra through "side effects" as
  each spectrum is saved, both as raw counts data and optionally as spectral
  irradiance or counts-per-second  spectral data in an \code{.rda} file as
  objects of the classes defined in package 'photobiology'. Optionally, the
  plot for each spectrum is saved as a \code{.pdf} file. At any time, the
  current group of spectra can be saved as a collection. When a collection
  is created, spectral data are for several spectra are saved together.
  Summaries are saved to a CSV file and joint plots to a \code{.pdf} file.
  The value returned by the function is that from closing the
  connection to the spectrometer.
}
\description{
Interactive front-end allowing acquisition of spectral irradiance using Ocean
Optics spectrometers. Output of spectral data in R data files stored in
objects suitable for use with packages 'photobiology' and 'ggspectra' as well
as plots as PDF files and summaries as comma separated files.
}
\details{
Function \code{acq_irrad_interactive()} supports measurement of
  spectral irradiance from continuous light sources. (It assumes that the
  duration of the measurement event is the relevant time base for expression
  of the flux of radiation.) This function can be used to acquire spectral
  irradiance using different protocols for acquisition and stray light and
  dark corrections. The protocols are described in the vignettes and in the
  help for the lower level functions called, also from this same package.

  Using this function only requires an Ocean Optics spectrometer to be
  connected to the computer where R is running and the OmniDriver runtime
  from Ocean Insight installed. The connection to the spectrometer and
  selection of channel, when relevant, is done from within these functions.

  The irradiance calibration will be retrieved from the spectrometer memory
  as a last resource if not supplied in any other way. Given that the factors
  are stored by Ocean Optics in a format that ignores the entrance optics,
  either the effective cosine diffuser area in xxx should be passed to
  parameter \code{area} or a character string with the type of the diffuser
  passed to \code{diff.type}. If no irradiance calibration is available,
  counts per second (cps) or raw counts are the only options available.

  Three default protocols are available by default. They differ in the
  additional measurements done to correct for stray light and dark noise. The
  default protocols are usually suitable, if new protocols are passed, each
  character vector must contain strings "light", "filter" and "dark".

  By default the integration time is set automatically so that the number of
  counts at the highest peak is close to 1 - \code{target.margin} times the
  maximum of the range of the instrument detector (retrieved from the
  calibration or the instrument memory). The minimum \code{tot.time} is
  obtained by increasing the number of scans. The maximum integration time
  supported by the spectrometer is not exceeded.

  Plots are produced with functions from package 'ggspectra' and respect the
  default annotations set with function \code{set_annotations_default()},
  default wavebands set with function \code{set_w.band_default()}, and
  irradiance quantities set with \code{photon_as_default()}, and
  \code{energy_as_default()}.

  The different interface modes available are suitable for different types of
  measurements.
}
\note{
Calibration data needs in most cases to be imported into R and
  parameters entered for the special correction algorithms into a correction
  method descriptor. The corrections are skipped if the needed information is
  missing. If no spectral irradiance calibration is available and attempt is
  made to retrieve it from the spectrometer, but given the format used by
  Ocean Optics/Ocean Insight, in this case the effective \code{area} of the
  cosine diffuser used (or the model name if from Ocean Optics) should be
  supplied by the user.

  The function is composed in a modular way from functions that can be
  reshuffled and combined with other functions to define new variations
  possibly better suited to users' needs and tastes.
}
\examples{
# please, see also the example scripts installed with the package

\dontrun{
# requires an Ocean Optics spectrometer to be connected via USB

acq_irrad_interactive()
acq_irrad_interactive(qty.out = "cps")

}

}
\seealso{
This function calls functions \code{\link{tune_interactive}},
  \code{\link{protocol_interactive}} and
  \code{\link{set_attributes_interactive}}. If irradiance calibration is
  retrieved from the instrument, functions \code{\link{get_oo_descriptor}}
  and \code{\link{oo_calib2irrad_mult}} are also called.

Other interactive acquisition functions: 
\code{\link{acq_fluence_interactive}()},
\code{\link{acq_fraction_interactive}()}
}
\concept{interactive acquisition functions}
