% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.spc.obj.R
\name{acq_raw_mspct}
\alias{acq_raw_mspct}
\title{Take one set of spectral readings}
\usage{
acq_raw_mspct(oo_descriptor, acq_settings, protocol = c("measure", "filter",
  "dark"), user.label, geocode = NA, verbose = TRUE)
}
\arguments{
\item{protocol}{a vector of character strings}

\item{verbose}{a logical to enable or disable warnings}

\item{x}{an initialized list of raw spectra containing parameter values}

\item{add.label}{a character string to be concatenated to the current label}
}
\value{
a raw_mspct object
}
\description{
Take readings according to parameters from a list of settings and a protocol
defined by a vector of names.
}
\keyword{manip}
\keyword{misc}

