% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acquire.R
\name{acq_raw_mspct}
\alias{acq_raw_mspct}
\title{Take one set of spectral readings}
\usage{
acq_raw_mspct(oo_descriptor, acq_settings, protocol = c("measure", "filter",
  "dark"), user.label, geocode = data.frame(lon = NA_real_, lat = NA_real_),
  verbose = TRUE)
}
\arguments{
\item{oo_descriptor}{list as returned by function \code{get_oo_descriptor()}}

\item{acq_settings}{list as returned by functions \code{tune_acq_settings()}
or \code{retune_acq_settings()} or \code{acq_settings()}}

\item{protocol}{vector of character strings}

\item{user.label}{character string to set as label}

\item{geocode}{data.frame with at least columns "lon" and "lat" compatible
with value returned by \code{ggmap::geocode()}}

\item{verbose}{ogical to enable or disable warnings}
}
\value{
a raw_mspct object
}
\description{
Take readings according to parameters from a list of settings and a protocol
defined by a vector of names.
}

