% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-raw.R
\name{acq_raw_spct}
\alias{acq_raw_spct}
\title{Measure one raw spectrum}
\usage{
acq_raw_spct(descriptor, acq.settings, num.exposures = -1L,
  f.trigger.pulses = f.trigger.message, what.measured = NA,
  where.measured = data.frame(lon = NA_real_, lat = NA_real_),
  set.all = TRUE, verbose = TRUE)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{acq.settings}{list as returned by functions \code{tune_acq_settings}}

\item{num.exposures}{integer Number or light pulses (flashes) per scan. Set
to \code{-1L} to indicate that the light source is continuous.}

\item{f.trigger.pulses}{function Function to be called to trigger light
pulse(s). Should accept as its only argument the number of pulses, and
return \code{TRUE} on sucess and \code{FALSE} on failure.}

\item{what.measured}{value used to set attribute}

\item{where.measured}{data.frame with at least columns "lon" and "lat"
compatible with value returned by \code{ggmap::geocode()}}

\item{set.all}{logical resend or not all instrument settings}

\item{verbose}{logical to enable or disable warnings}
}
\value{
a "raw_spct" object with one or more columns containing raw counts
and one column with wavelength data.
}
\description{
Take one spectral measurement which depending on the settings can consist
on more than one raw spectrum meant to represent a SINGLE observation after
conversion into calibrated data such as in the case of HDR.
}
