% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-raw.R
\name{acq_raw_spct}
\alias{acq_raw_spct}
\title{Measure one raw spectrum}
\usage{
acq_raw_spct(
  descriptor,
  acq.settings,
  f.trigger.pulses = f.trigger.message,
  what.measured = NA,
  where.measured = data.frame(lon = NA_real_, lat = NA_real_),
  set.all = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}.}

\item{acq.settings}{list as returned by functions \code{tune_acq_settings}.}

\item{f.trigger.pulses}{function Function to be called to trigger light
pulse(s). Should accept as its only argument the number of pulses, and
return \code{TRUE} on sucess and \code{FALSE} on failure.}

\item{what.measured}{value used to set attribute.}

\item{where.measured}{data.frame with at least columns "lon" and "lat"
compatible with value returned by \code{ggmap::geocode()}.}

\item{set.all}{logical resend or not all instrument settings.}

\item{verbose}{logical to enable or disable warnings.}
}
\value{
A \code{raw_spct} object with one column \code{w.length} and one
  column \code{counts}, or two or more columns \code{counts1, counts2, ...}
  containing raw counts data. The number of columns with raw counts is
  determined by \code{acq.settings}, with multiple columns in the case of
  integration time bracketing or HDR.
}
\description{
Take one spectral measurement which depending on the settings can consist
in multiple raw spectra meant to represent a SINGLE observation after
conversion into calibrated data, such as in the case of bracketing of
integration time for HDR.
}
\seealso{
\code{\link{acq_raw_mspct}} which can be used to acquire multiple
spectra according to a user defined protocol.

Other raw-counts-spectra acquisition functions: 
\code{\link{acq_raw_mspct}()},
\code{\link{hs_acq_raw_mspct}()}
}
\concept{raw-counts-spectra acquisition functions}
