% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.spc.obj.R
\name{acq_raw_spct}
\alias{acq_raw_spct}
\title{Measure one raw spectrum}
\usage{
acq_raw_spct(oo_descriptor, acq_settings, what.measured = NA,
  verbose = TRUE)
}
\arguments{
\item{oo_descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{acq_settings}{list as returned by functions \code{tune_acq_settings}}

\item{what.measured}{value used to set attribute}

\item{verbose}{ogical to enable or disable warnings}
}
\value{
a "raw_spct" object with one or more columns containing raw counts
and one column with wavelength data.
}
\description{
Take one spectral measurement which depending on the settings can consist
on more than one raw spectrum meant to represent a SINGLE observation after
conversion into calibrated data such as in the case of HDR.
}

