% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-settings.R
\name{acq_settings}
\alias{acq_settings}
\alias{tune_acq_settings}
\title{Settings for spectral measurement}
\usage{
acq_settings(oo_descriptor, integ.time, num.scans = 10L,
  start.integ.time = 0.05, min.integ.time = -Inf, max.integ.time = Inf,
  min.tot.time = 2, HDR.mult = c(short = 1, long = 10),
  pix.selector = TRUE, verbose = TRUE)

tune_acq_settings(oo_descriptor, acq_settings, verbose = TRUE)

tune_acq_settings(oo_descriptor, acq_settings, verbose = TRUE)
}
\arguments{
\item{oo_descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{integ.time}{numeric vaue in seconds}

\item{num.scans}{integer}

\item{start.integ.time}{numeric vaue in seconds}

\item{min.integ.time}{numeric vaue in seconds}

\item{max.integ.time}{numeric vaue in seconds}

\item{min.tot.time}{numeric vaue in seconds}

\item{HDR.mult}{a numeric vector with integ.time multipliers to be used for
"bracketing".}

\item{pix.selector}{a logical or numeric vector used as subscript to select
pixels}

\item{verbose}{a logical to enable or disable warnings}

\item{acq_settings}{list as returned by a previous call to \code{acq_settings()},
or \code{tune_acq_settings()}.}
}
\value{
a list.

a list.

a list.
}
\description{
Validate parameters for spectral measurements and return a list of values
usable as input for functions \code{retune_acq_settings()}, \code{acq_sptc()},
and \code{acq_mspct()}.

Find parameters for spectral measurements with a given measurements
protocol.

Find optimal settings for spectral measurements under a given measurement
protocol.
}
\details{
\code{acq_settings()} is used to manually save a complete set of
  instrument settings in a way that they can be reused as needed for repated
  acquisition of spectra. \code{tune_acq_settings()} and
  \code{tune_acq_settings()} differ only in their formal parameters and
  optimize settings to the current spectral irradiance.
}
\note{
\code{pixel.selector} can be used for two different purposes: to
  ignore bad pixels and to restrict integration-time tuning to the response
  from a range of pixels.

Ocean Optics spectrometers can be queried for the maximum and minimum
  supported integration times. This function modifies the user supplied
  values if outside these bounds. The defaults of -Inf and Inf force the use
  of the whole valid range of integration time supported by the connected
  intrument. \code{pixel.selector} can be used for two different purposes: to
  ignore bad pixels and to restrict integration-time tuning to the response
  from a range of pixels.
}

