% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-settings.R
\name{acq_settings}
\alias{acq_settings}
\alias{tune_acq_settings}
\title{Settings for spectral measurement}
\usage{
acq_settings(
  descriptor,
  integ.time = 0.01,
  num.scans = 10L,
  min.integ.time = -Inf,
  max.integ.time = Inf,
  tot.time.range = c(0, Inf),
  HDR.mult = ifelse(any(num.exposures != -1L), rep(1, length(num.exposures)), c(short =
    1, long = 10)),
  target.margin = 0.1,
  pix.selector = TRUE,
  corr.elect.dark = 0L,
  corr.sensor.nl = 0L,
  boxcar.width = 0L,
  force.valid = FALSE,
  num.exposures = -1L,
  verbose = TRUE
)

tune_acq_settings(descriptor, acq.settings, verbose = TRUE)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{integ.time}{numeric vaue in seconds}

\item{num.scans}{integer}

\item{min.integ.time}{numeric vaue in seconds}

\item{max.integ.time}{numeric vaue in seconds}

\item{tot.time.range}{numeric vector of length two with values in seconds}

\item{HDR.mult}{a numeric vector with integ.time multipliers to be used for
"bracketing".}

\item{target.margin}{numeric (0..1)}

\item{pix.selector}{a logical or numeric vector used as subscript to select
pixels}

\item{corr.elect.dark, corr.sensor.nl}{integer 0L (FALSE) or 1L (TRUE)}

\item{boxcar.width}{integer Number of pixels to average}

\item{force.valid}{logical Accept all spectra as valid, for example do not
treat clipping as an error condition.}

\item{num.exposures}{integer Number of flashes triggered per scan.}

\item{verbose}{a logical to enable or disable warnings}

\item{acq.settings}{list as returned by a previous call to
\code{acq_settings()}, or \code{tune_acq_settings()}.}
}
\value{
a list.

a list Containing the tuned settings.
}
\description{
Validate parameters for spectral measurements and return a list of values
usable as input for functions \code{retune_acq_settings()}, \code{acq_sptc()},
and \code{acq_mspct()}.

Find optimal settings for spectral measurements under a given measurement
protocol.
}
\details{
\code{acq_settings()} is used to create a complete set of
  instrument settings in a way that they can be reused as needed for repated
  acquisition of spectra. \code{acq_settings()} is an object constructor and
  \code{tune_acq_settings()} takes as argument a stored object containing
  settings and tunes them to optimize them for the measurement of the current
  spectral irradiance.

This function searches for the optimal integration time for a
  given condition by trial and error helped by interpolation and
  extrapolation when readings are not saturated. In the case of clipping or
  sensor signal saturation the integration time is decreased until clipping
  is avoided and then it is increased until optimal.
}
\note{
\code{pixel.selector} can be used for two different purposes: to
  ignore bad pixels and to restrict integration-time tuning to the response
  from a range of pixels. The interpretation of \code{tot.time.range} is
  as follows: first value is minimum time, second value is maximum time.
  If both values are the same, then an exact measurement time is computed.

Ocean Optics spectrometers can be queried for the maximum and minimum
  supported integration times. This function modifies the user supplied
  values if outside these bounds. The defaults of -Inf and Inf force the use
  of the whole valid range of integration time supported by the connected
  intrument. \code{pixel.selector} can be used for two different purposes: to
  ignore bad pixels and to restrict integration-time tuning to the response
  from a range of pixels.

The returned value can be used as argument to \code{acq.settings} in
  other functions like \code{\link{acq_raw_spct}} and
  \code{\link{acq_raw_mspct}}
}
\seealso{
Other acquisition-settings related functions: 
\code{\link{set_integ_time}()},
\code{\link{set_linearized}()},
\code{\link{set_num_exposures}()}
}
\concept{acquisition-settings related functions}
