% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-raw-data.R
\docType{data}
\name{blue_filter.raw_mspct}
\alias{blue_filter.raw_mspct}
\title{Raw counts data for a filter measurement}
\format{A raw_mscpt}
\usage{
blue_filter.raw_mspct
}
\description{
A blue interference filter from UQG measured using a household incandescent
lamp as light source. The spectrometer used was an Ocean Optics Maya2000 Pro.
}
\keyword{datasets}
