% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{choose_ch_interactive}
\alias{choose_ch_interactive}
\title{Interactively select a channel}
\usage{
choose_ch_interactive(instruments, sr.index = 0L,
  prompt.text = "Channels available: ")
}
\arguments{
\item{instruments}{the returm value of \code{list_instruments(w)}}

\item{sr.index}{integer The index to the spectrometer, starting from zero,
following Omni Driver indexing conventions.}

\item{prompt.text}{character string to use as prompt.}
}
\value{
A numeric index suitable for use in calls to functions defined in
package 'rOmniDriver' based on which package 'ooacquire' is coded.
}
\description{
Choice of channel to be used if spectrometer has more than one channel.
}
\seealso{
Other interactive acquisition utility functions: \code{\link{choose_sr_interactive}},
  \code{\link{f.trigger.message}},
  \code{\link{list_srs_interactive}},
  \code{\link{protocol_interactive}},
  \code{\link{set_attributes_interactive}},
  \code{\link{set_folder_interactive}},
  \code{\link{set_seq_interactive}},
  \code{\link{tune_interactive}}
}
\concept{interactive acquisition utility functions}
