% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{choose_sr_interactive}
\alias{choose_sr_interactive}
\title{Interactively select an instrument}
\usage{
choose_sr_interactive(instruments)
}
\arguments{
\item{instruments}{the returm value of \code{list_instruments(w)}.}
}
\description{
Choice of spectrometer by name from a list of serial numbers, allowing the
user to correct the selection if needed.
}
