% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-utils.R
\name{collect_spct_files}
\alias{collect_spct_files}
\title{Collect spectra into a collection}
\usage{
collect_spct_files(path = ".", range = NULL, strength = 0,
  name.root = path)
}
\arguments{
\item{path}{a character string giving
the name of the folder from wich to load .Rda files containing spectra.}

\item{range}{a numeric vector of length two, or any other object for which
function range() will return two.}

\item{strength}{numeric value to adjust the degree of smoothing.}

\item{name.root}{character string to prepend to the name of the spectra in
the collection.}
}
\description{
Read .Rda files as saved during data acquisition and build a collection with
the spectra read, possibly first trimming the range of wavelengths and/or
smoothing the spectral data.
}
\seealso{
Function \code{\link[photobiology]{trim_wl}} is used to trim the
  range of the data to plot, and function
  \code{\link[photobiology]{smooth_spct}} with \code{method = "supsmu"} are
  used in the implementation, and prameters \code{range} and \code{strength}
  are passed to them, respectively.
}
