% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.close.R
\name{end_session}
\alias{end_session}
\title{Disconnect from spectrometer}
\usage{
end_session(oo.descriptor)
}
\arguments{
\item{x}{a list as returned by \code{start_session()} or}
}
\description{
Close the connection to the spectrometer pointed
at by the Java object referenced in \code{x}.
}

