% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{end_session}
\alias{end_session}
\title{Disconnect from spectrometer}
\usage{
end_session(w)
}
\arguments{
\item{w}{a java wrapper as returned by \code{start_session()}}
}
\description{
Close the connection to the spectrometer pointed
at by the Java object referenced in \code{x}.
}
\seealso{
Other spectrometer-connection functions: \code{\link{list_instruments}},
  \code{\link{start_session}}
}
\concept{spectrometer-connection functions}
