% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{f.trigger.message}
\alias{f.trigger.message}
\title{Manual trigger pulses request}
\usage{
f.trigger.message(n = 1L)
}
\arguments{
\item{n}{integer Number of pulses (flashes) to trigger per call.}
}
\description{
This function is used by default. It prints a message asking the operator
to manually trigger the flash. A more elaborate function, using specific
hardware can be used to automatically trigger the light source, or to
enable hard triggering of the light source by the spectrometer itself.
}
\note{
When using this function, use an integration time that gives enough
time for the manual triggering of the flash reliably within the integration.
}
