% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-pulsed-interactive.R
\name{f.trigger.message}
\alias{f.trigger.message}
\title{Manual trigger request}
\usage{
f.trigger.message()
}
\description{
This function is used by default. It prints a message asking the operator
to manually trigger the flash.
}
