% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flame-s-data.R
\docType{data}
\name{FLMS00673_descriptors}
\alias{FLMS00673_descriptors}
\title{Instrument descriptors for a Flame S spectrometer}
\format{A list.}
\usage{
FLMS00673_descriptors
}
\description{
This a list of instrument descriptors, each one containing the calibration
data needed to read raw data acquired on different dates.
}
\keyword{datasets}
