% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{get_oo_descriptor}
\alias{get_oo_descriptor}
\title{Get the instrument description and EEPROM data}
\usage{
get_oo_descriptor(w, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{w}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer for the time being not
exported}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a list
}
\description{
Model, configuration, serial number, and calibration data stored in the
EEPROM of an Ocean Optics spectrometer are retrieved and returned in a
list.
}

