% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.spc.obj.R
\name{get_oo_settings}
\alias{get_oo_settings}
\title{Get the current values of instrument settings}
\usage{
get_oo_settings(w, sr.index = 0L, ch.index = 0L)
}
\arguments{
\item{w}{an open Wrapper object from Omnidriver}

\item{sr.index}{an index to address the spectrometer for the time being not
exported}

\item{ch.index}{an index to address the channel in a spectrometer with more
than one channel.}
}
\value{
a list
}
\description{
Query the spectrometer for the settings currently in use for corrections,
smotthing and acquisition parameters integration time and number of scans.
}

