% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.spc.obj.R
\name{get_oo_settings}
\alias{get_oo_settings}
\title{Replace wavelength values in an instrument description}
\usage{
get_oo_settings(oo_descriptor)
}
\arguments{
\item{oo_descriptor}{list as returned by function \code{get_oo_descriptor}}
}
\value{
a list
}
\description{
Get the current values of instrument settings
}
\details{
Query the spectrometer for the settings currently in use for corrections,
smotthing and acquisition parameters integration time and number of scans.
}

