% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{get_oo_settings}
\alias{get_oo_settings}
\title{Get the current values of instrument settings}
\usage{
get_oo_settings(descriptor)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}
}
\value{
a list
}
\description{
Query the spectrometer for the settings currently in use for corrections,
smotthing and acquisition parameters integration time and number of scans.
}
