% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-raw-data.R
\docType{data}
\name{halogen.raw_mspct}
\alias{halogen.raw_mspct}
\title{Raw counts data for a lamp measurement.}
\format{A raw_mscpt}
\usage{
halogen.raw_mspct
}
\description{
A household Osram tungsten halogen 25W E27 lamp measured at short distance. The
spectrometer used was an Ocean Optics Maya2000 Pro.
}
\keyword{datasets}
