% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hs-acq-raw.R
\name{hs_acq_raw_mspct}
\alias{hs_acq_raw_mspct}
\title{Acquire spectra at high speed}
\usage{
hs_acq_raw_mspct(
  descriptor,
  acq.settings,
  num.spectra = 100L,
  base.name = NULL,
  f.trigger.on = f.trigger.message,
  f.trigger.off = NA,
  what.measured = NA,
  where.measured = data.frame(lon = NA_real_, lat = NA_real_),
  set.all = TRUE,
  verbose = TRUE,
  return.list = FALSE
)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}.}

\item{acq.settings}{list as returned by functions \code{tune_acq_settings}.}

\item{num.spectra}{integer Number of individual spectra to acquire.}

\item{base.name}{character The name given to individual spectra is formed
by this string followed by a sequential numeric index.}

\item{f.trigger.on, f.trigger.off}{function Functions to be called
immediately before and immediately after a measurement. See
\code{\link{acq_raw_spct}} for details.}

\item{what.measured}{value used to set attribute.}

\item{where.measured}{data.frame with at least columns "lon" and "lat"
compatible with value returned by \code{ggmap::geocode()}.}

\item{set.all}{logical resend or not all instrument settings.}

\item{verbose}{logical to enable or disable warnings.}

\item{return.list}{logical Return a \code{list} instead of a \code{raw_mspct}
object.}
}
\value{
A \code{raw_mspct} containing one \code{raw_spct} object with one
  column \code{w.length} and one column \code{counts} for each spectrum. The
  number of columns with raw counts is always one and integration time
  bracketing or HDR values in \code{acq.settings} are ignored except for the
  smallest value.
}
\description{
Take one set of spectra at high speed using unchanged instrument settings
special OmniDriver API functions for buffered acquisition. No HDR bracketing
is possible, and synchronization with a pulsed light source is not supported.
}
\seealso{
For normal speed acquisition of a single spectrum and multiple
  spectra according to a user defined protocol see \code{\link{acq_raw_spct}}
  and \code{\link{acq_raw_mspct}}.

Other raw-counts-spectra acquisition functions: 
\code{\link{acq_raw_mspct}()},
\code{\link{acq_raw_spct}()}
}
\concept{raw-counts-spectra acquisition functions}
