% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instr-defaults.R
\name{instr_defaults}
\alias{instr_defaults}
\title{Fetch spectrometer defaults}
\usage{
instr_defaults(
  serial_no,
  entrance.optics = "cosine",
  date = lubridate::now(tzone = "UTC"),
  descriptors = NA,
  correction.method = NA
)
}
\arguments{
\item{serial_no}{character The serial number of an Ocean Optics spectrometer
for which data is available in 'ooacquire'.}

\item{entrance.optics}{character, name or geometry of diffuser, needed only
if there is more than one for the same instrument.}

\item{date}{Any object that \code{anytime::anydate()} will decode as a date
or convert to a date. Used to select a descriptor containing calibration
data valid for the date when \emph{data was acquired}.}

\item{descriptors}{list A list of instrument descriptors containing
calibration data.}

\item{correction.method}{list The method to use when applying the calibration}
}
\description{
Find default instrument descriptor and correction method based on
spectrometer serial number and entrance optics.
}
\details{
When \code{protocols} or \code{descriptor} is not available
  (\code{NA}), the missing value is replaced by a default based on the
  arguments passed to \code{serial_no}, \code{entrance.optics}, and
  \code{protocols}.
}
\examples{

str(instr_defaults(serial_no = "MAYP114590"))

}
