% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaz-data.R
\docType{data}
\name{JAZA3098_calib_dates.df}
\alias{JAZA3098_calib_dates.df}
\alias{JAZA3098_descriptors}
\alias{JAZA3098_ch1_descriptors}
\alias{JAZA3098_ch1_ylianttila.mthd}
\alias{JAZA3098_ch1_sun.mthd}
\alias{JAZA3098_ch1_simple.mthd}
\alias{JAZA3098_ch1_none.mthd}
\alias{JAZA3098_ch2_descriptors}
\alias{JAZA3098_ch2_ylianttila.mthd}
\alias{JAZA3098_ch2_sun.mthd}
\alias{JAZA3098_ch2_simple.mthd}
\alias{JAZA3098_ch2_none.mthd}
\title{Jaz spectrometer s/n JAZA3098}
\format{
Lists and data frames.

An object of class \code{list} of length 1.

An object of class \code{list} of length 0.

An object of class \code{list} of length 0.

An object of class \code{list} of length 9.

An object of class \code{list} of length 9.

An object of class \code{list} of length 1.

An object of class \code{list} of length 0.

An object of class \code{list} of length 0.

An object of class \code{list} of length 9.

An object of class \code{list} of length 9.
}
\usage{
JAZA3098_calib_dates.df

JAZA3098_ch1_descriptors

JAZA3098_ch1_ylianttila.mthd

JAZA3098_ch1_sun.mthd

JAZA3098_ch1_simple.mthd

JAZA3098_ch1_none.mthd

JAZA3098_ch2_descriptors

JAZA3098_ch2_ylianttila.mthd

JAZA3098_ch2_sun.mthd

JAZA3098_ch2_simple.mthd

JAZA3098_ch2_none.mthd
}
\description{
Calibration data, descriptors and methods for correction of stray light and
high dynamic range based on multiple integrations times per spectrum.
}
\note{
Some of the method definitions are fill-ins for methods not implemented.
In this case no calibration for spectral irradiance or characterizations of
the slit function are available.
}
\seealso{
Other objects containing instrument-specific data: 
\code{\link{FLMS00416_descriptors}},
\code{\link{FLMS00440_descriptors}},
\code{\link{FLMS00673_descriptors}},
\code{\link{FLMS04133_descriptors}},
\code{\link{MAYP112785_descriptors}},
\code{\link{MAYP11278_descriptors}},
\code{\link{MAYP114590_descriptors}}
}
\concept{objects containing instrument-specific data}
\keyword{datasets}
