% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{list_instruments}
\alias{list_instruments}
\title{List connected spectrometers}
\usage{
list_instruments(w)
}
\arguments{
\item{w}{an open Wrapper object from Omnidriver}
}
\value{
a list
}
\description{
List all connected spectrometers by model name and serial number, plus
the number of channels in each of them.
}
\seealso{
Other spectrometer-connection functions: \code{\link{end_session}},
  \code{\link{start_session}}
}
\concept{spectrometer-connection functions}
