% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{list_srs_interactive}
\alias{list_srs_interactive}
\title{Get list of conencted instruments}
\usage{
list_srs_interactive(w)
}
\arguments{
\item{w}{handle to Omni Driver, used to test if an spectrometer is still
connected.}
}
\description{
Get list of spectrometers requesting user to connect one or abort if none
fould.
}
\seealso{
Other interactive acquisition utility functions: 
\code{\link{choose_ch_interactive}()},
\code{\link{choose_sr_interactive}()},
\code{\link{f.trigger.message}()},
\code{\link{protocol_interactive}()},
\code{\link{set_attributes_interactive}()},
\code{\link{set_folder_interactive}()},
\code{\link{set_seq_interactive}()},
\code{\link{tune_interactive}()}
}
\concept{interactive acquisition utility functions}
