% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-oo-header.R
\name{map_oofile_header_rows}
\alias{map_oofile_header_rows}
\title{Set the instrument description.}
\usage{
map_oofile_header_rows(lines, header.end = NULL, grammar = oo.minimum.gr)
}
\arguments{
\item{lines}{character The data file read line by line.}

\item{header.end}{integer The index to the last line of the longest
header expected.}

\item{grammar}{data.frame With character variables "feature" and "pattern"
that will be used to locate the lines containing each type of metadata.}
}
\value{
A data frame with two variables, "feature" of class character, and
  "line.idx" of class integer.
}
\description{
Locate in which lines of the file header different metadata features are
  located.
}
