% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP112785_descriptors}
\alias{MAYP112785_descriptors}
\alias{MAYP112785_calib_dates.df}
\alias{MAYP112785_ylianttila.mthd}
\alias{MAYP112785_sun.mthd}
\alias{MAYP112785_simple.mthd}
\title{Maya2000 Pro  spectrometer s/n MAYP112785}
\format{
Lists and data frames.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 4 rows and 8 columns.

An object of class \code{list} of length 10.

An object of class \code{list} of length 10.

An object of class \code{list} of length 10.
}
\usage{
MAYP112785_descriptors

MAYP112785_calib_dates.df

MAYP112785_ylianttila.mthd

MAYP112785_sun.mthd

MAYP112785_simple.mthd
}
\description{
List of instrument descriptors, each one containing the calibration
data needed to convert raw data acquired on different dates. Calibration
data from STUK's non-ionising radiation laboratory. Correction method
algorithms for stray light and slit function by Lasse Ylianttila, and
variations by Pedro J. Aphalo..
}
\seealso{
Other objects containing instrument-specific data: 
\code{\link{FLMS00416_descriptors}},
\code{\link{FLMS00440_descriptors}},
\code{\link{FLMS00673_descriptors}},
\code{\link{FLMS04133_descriptors}},
\code{\link{JAZA3098_calib_dates.df}},
\code{\link{MAYP11278_descriptors}},
\code{\link{MAYP114590_descriptors}}
}
\concept{objects containing instrument-specific data}
\keyword{datasets}
