% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP112785_descriptors}
\alias{MAYP112785_descriptors}
\title{Instrument descriptors for a Maya2000 Pro  spectrometer}
\format{A list.}
\usage{
MAYP112785_descriptors
}
\description{
This a list of instrument descriptors, each one containing the calibration
data needed to read raw data acquired on different dates.
}
\keyword{datasets}

