% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP112785_descriptors}
\alias{MAYP112785_descriptors}
\title{Instrument descriptors for a Maya2000 Pro  spectrometer}
\format{A list.}
\usage{
MAYP112785_descriptors
}
\description{
This a list of instrument descriptors, each one containing the calibration
data needed to read raw data acquired on different dates.
}
\seealso{
Other objects containg instrument-specific data: \code{\link{FLMS00416_cal.spct}},
  \code{\link{FLMS00416_calib_dates.df}},
  \code{\link{FLMS00416_descriptors}},
  \code{\link{FLMS00416_none.mthd}},
  \code{\link{FLMS00416_simple.mthd}},
  \code{\link{FLMS00416_sun.mthd}},
  \code{\link{FLMS00416_ylianttila.mthd}},
  \code{\link{FLMS00440_cal.spct}},
  \code{\link{FLMS00440_descriptors}},
  \code{\link{FLMS00440_none.mthd}},
  \code{\link{FLMS00440_simple.mthd}},
  \code{\link{FLMS00440_sun.mthd}},
  \code{\link{FLMS00440_ylianttila.mthd}},
  \code{\link{FLMS00673_cal.spct}},
  \code{\link{FLMS00673_calib_dates.df}},
  \code{\link{FLMS00673_descriptors}},
  \code{\link{FLMS00673_none.mthd}},
  \code{\link{FLMS00673_simple.mthd}},
  \code{\link{FLMS00673_sun.mthd}},
  \code{\link{FLMS00673_ylianttila.mthd}},
  \code{\link{FLMS04133_cal.spct}},
  \code{\link{FLMS04133_calib_dates.df}},
  \code{\link{FLMS04133_descriptors}},
  \code{\link{FLMS04133_none.mthd}},
  \code{\link{FLMS04133_simple.mthd}},
  \code{\link{FLMS04133_sun.mthd}},
  \code{\link{FLMS04133_ylianttila.mthd}},
  \code{\link{JAZA3098_calib_dates.df}},
  \code{\link{JAZA3098_ch1_descriptors}},
  \code{\link{JAZA3098_ch1_simple.mthd}},
  \code{\link{JAZA3098_ch1_sun.mthd}},
  \code{\link{JAZA3098_ch1_ylianttila.mthd}},
  \code{\link{JAZA3098_ch2_descriptors}},
  \code{\link{JAZA3098_ch2_none.mthd}},
  \code{\link{JAZA3098_ch2_simple.mthd}},
  \code{\link{JAZA3098_ch2_sun.mthd}},
  \code{\link{JAZA3098_ch2_ylianttila.mthd}},
  \code{\link{MAYP112785_calib_dates.df}},
  \code{\link{MAYP112785_simple.mthd}},
  \code{\link{MAYP112785_sun.mthd}},
  \code{\link{MAYP112785_ylianttila.mthd}},
  \code{\link{MAYP11278_calib_dates.df}},
  \code{\link{MAYP11278_descriptors}},
  \code{\link{MAYP11278_simple.mthd}},
  \code{\link{MAYP11278_sun.mthd}},
  \code{\link{MAYP11278_ylianttila.mthd}}
}
\concept{objects containg instrument-specific data}
\keyword{datasets}
