% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MAYP112785_tail_correction}
\alias{MAYP112785_tail_correction}
\title{Function to compute the tail correction}
\usage{
MAYP112785_tail_correction(w_length, cts_second)
}
\arguments{
\item{w_length}{numeric vector of wavelengths in nm.}

\item{cts_second}{numeric vector of counts per second (stray-light corrected).}
}
\value{
tail returned as numeric vector within a list.

Tail correction is a reimplementation of the calculations developed by Lasse
Ylianttila (STUK, Finland), originally in Excel.
}
\description{
Function to compute the tail correction
}
\seealso{
Other spectral data-processing functions: 
\code{\link{MAYP11278_tail_correction}()},
\code{\link{check_sn_match}()},
\code{\link{linearize_counts}()},
\code{\link{merge_raw_mspct}()},
\code{\link{new_correction_method}()},
\code{\link{ref_correction}()},
\code{\link{trim_counts}()},
\code{\link{uvb_corrections}()}
}
\concept{spectral data-processing functions}
