% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP114590_descriptors}
\alias{MAYP114590_descriptors}
\alias{MAYP114590_cal.spct}
\alias{MAYP114590_calib_dates.df}
\alias{MAYP114590_ylianttila.mthd}
\alias{MAYP114590_sun.mthd}
\alias{MAYP114590_simple.mthd}
\alias{MAYP114590_none.mthd}
\title{Maya2000 Pro spectrometer s/n MAYP114590}
\format{
Lists and data frames.

An object of class \code{calibration_spct} (inherits from \code{generic_spct}, \code{spec_tbl_df}, \code{tbl_df}, \code{tbl}, \code{data.frame}) with 2068 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1 rows and 7 columns.

An object of class \code{list} of length 0.

An object of class \code{list} of length 0.

An object of class \code{list} of length 12.

An object of class \code{list} of length 12.
}
\usage{
MAYP114590_descriptors

MAYP114590_cal.spct

MAYP114590_calib_dates.df

MAYP114590_ylianttila.mthd

MAYP114590_sun.mthd

MAYP114590_simple.mthd

MAYP114590_none.mthd
}
\description{
List of instrument descriptors, each one containing the calibration
data needed to convert raw data acquired on different dates. Calibration
data from STUK's non-ionising radiation laboratory. Correction method
algorithms for stray light and slit function by Lasse Ylianttila, and
variations by Pedro J. Aphalo..
}
\note{
Some of the method definitions are fill-ins for methods not implemented.
In this case no characterization of the slit function is available.
}
\seealso{
Other objects containing instrument-specific data:
\code{\link{FLMS00416_descriptors}},
\code{\link{FLMS00440_descriptors}},
\code{\link{FLMS00673_descriptors}},
\code{\link{FLMS04133_descriptors}},
\code{\link{JAZA3098_calib_dates.df}},
\code{\link{MAYP112785_descriptors}},
\code{\link{MAYP11278_descriptors}}
}
\concept{objects containing instrument-specific data}
\keyword{datasets}
