% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maya-data.R
\docType{data}
\name{MAYP114590_descriptors}
\alias{MAYP114590_descriptors}
\title{Instrument descriptors for a Maya2000 Pro  spectrometer}
\format{
A list.
}
\usage{
MAYP114590_descriptors
}
\description{
List of instrument descriptors, each one containing the calibration
data needed to convert raw data acquired on different dates. Calibration data
from Ocean Optics (currently Ocean Insight).
}
\seealso{
Other objects containing instrument-specific data: 
\code{\link{MAYP112785_calib_dates.df}},
\code{\link{MAYP112785_descriptors}},
\code{\link{MAYP112785_simple.mthd}},
\code{\link{MAYP112785_sun.mthd}},
\code{\link{MAYP112785_ylianttila.mthd}},
\code{\link{MAYP11278_calib_dates.df}},
\code{\link{MAYP11278_descriptors}},
\code{\link{MAYP11278_simple.mthd}},
\code{\link{MAYP11278_sun.mthd}},
\code{\link{MAYP11278_ylianttila.mthd}},
\code{\link{MAYP114590_cal.spct}},
\code{\link{MAYP114590_calib_dates.df}},
\code{\link{MAYP114590_none.mthd}},
\code{\link{MAYP114590_simple.mthd}},
\code{\link{MAYP114590_sun.mthd}},
\code{\link{MAYP114590_ylianttila.mthd}}
}
\concept{objects containing instrument-specific data}
\keyword{datasets}
