% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim-counts.R
\name{merge_cps}
\alias{merge_cps}
\title{Merge counts per second data}
\usage{
merge_cps(x)
}
\arguments{
\item{x}{cps_spct object}
}
\value{
a copy of x with values replaced as needed in all counts columns
  present.
}
\description{
Replace NAs in longer integration time cps with cps values from shorter
integration.
}

