% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-R.R
\name{new_correction_method}
\alias{new_correction_method}
\title{Make default method from descriptor}
\usage{
new_correction_method(descriptor, stray.light.method = "none",
  stray.light.wl = c(220, 230), flt.dark.wl = c(200, 210),
  flt.ref.wl = c(360, 379.5), flt.Tfr = 1)
}
\arguments{
\item{descriptor}{list, as returned by \code{get_oo_descriptr()}}

\item{stray.light.method}{character Name of method.}

\item{stray.light.wl, flt.dark.wl, flt.ref.wl}{numeric vector with wavelengths (nm).}

\item{flt.Tfr}{numeric Transmittance of filter as a fraction (0...1).}
}
\value{
a list
}
\description{
A function that builds a default method from an instrument descriptor.
Useful when the spectrometers has not been characterized as needed for
the more sofisticated methods. Can use stray light correction
but not slit function correction. Stray light correction is valid
only if it was also used during irradiance callibration. Suitablity
of wavelengths and method depends on the instrument configuration.
}
\details{
The currently recognized methods for stray-light correction are
\code{"none"}, \code{"original"}, \code{"sun"}, and \code{"simple"}. With
the default method \code{"none"}, the values of the remaining parameters
are ignored.
}
\note{
Defaults for indexes are for the first channel of the first
  spectrometer currently connected.
}
