% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooacquire.R
\docType{package}
\name{ooacquire-package}
\alias{ooacquire}
\alias{ooacquire-package}
\title{ooacquire: Acquire Data from OO Spectrometers}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions to acquire data directly from Ocean Optics spectrometers, and functions to read similar data from files. Functions to convert raw-counts into counts-per-second and physical quantities. Data are saved in objects of classes defined in package 'photobiology'. The instrument settings, instrument description, date-time of acquisition and optionally goecode are stored as attributes.
}
\details{
Processing of raw-counts data acquired following different protocols and
corrections including integration-time bracketing and merging, slit function
corrections, subtraction of stray light and adaptive smoothing corrections.

Acquisition of raw-counts spectra directly from Ocean Optics' spectrometers.
A high level interface to the OmniDriver library for spectrometer
control and data acquisition provided by Ocean Optics for their
instruments. Built on top of the 'rOmniDriver' package which
provides a very thin wrapper on the Java functions giving access to
the proprietary driver.

Reading of raw-counts spectra from files saved by Ocean Optics' software:
OceanView, SpectraSuite, Raspberry Pi SDK, and Jaz firmware.
}
\note{
This package is provided \strong{without any warranty of suitability
  for any specific purpose or instrument} as it has been tested with only
  three different models of Ocean Optics spectrometers, each with one or two
  of many available hardware configurations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.r4photobiology.info/ooacquire/}
  \item \url{https://github.com/aphalo/ooacquire}
  \item Report bugs at \url{https://github.com/aphalo/ooacquire/issues}
}

}
\author{
\strong{Maintainer}: Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} [translator]

Authors:
\itemize{
  \item Pedro J. Aphalo \email{pedro.aphalo@helsinki.fi} [translator]
  \item Lasse Ylianttila
}

Other contributors:
\itemize{
  \item Titta K. Kotilainen [contributor]
}

}
