% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ooacquire.R
\docType{package}
\name{ooacquire-package}
\alias{ooacquire}
\alias{ooacquire-package}
\title{Acquire Spectral Data from Ocean Optics Spectrometers}
\description{
Functions to acquire data directly from spectrometers,
automatically adjusting integration time and numbers of scans to average.
Auxiliary functions to read the instrument description from its firmware ,
and to read current instrument settings. Function to apply linearization
function to data (when linearization is sisabled in the instrument).
Built as an extenssion to package 'photobiology'. Spectral data is returned
in objects of class "raw_spct" or "raw_mspct". The instrument settings,
instrument description, date-time of acquisition and optionally goecode are
stored as attributes.
}
\details{
A higher level interface to the OmniDriver library for spectrometer
control and data acquisition provided by Ocean Optics for their
instruments. It is built on top of the 'rOmniDriver' package which
provides a very thin wrapper on the Java functions giving access to
the proprietary driver.
}
\note{
This package is provided \strong{without nay warranty of suitability
  for any specific purpose or instrument} as it has been tested with only
  three different models of Ocean Optics spectrometers, each one with one of
  many possible configurations.
}

