% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-oo-header.R
\name{oofile_data_rows}
\alias{oofile_data_rows}
\title{Find range of lines in file containing spectral data.}
\usage{
oofile_data_rows(lines)
}
\arguments{
\item{lines}{character The data file read line by line.}
}
\value{
integer vector of length two, containing the indexes to the first
  last rows containing spectral data.
}
\description{
Ocean Optics files can have varying numbers of rows containing spectral
  data, but these rows are fenced between two lines containing recognizable
  character strings. This function, detects these strings.
}
\note{
This is a rather slow operation, so if all files to be read have
  the same format, it is inefficient to call this function for each file.
}
\seealso{
Other functions for importing spectral data from files: \code{\link{collect_spct_files}},
  \code{\link{map_oofile_header_rows}},
  \code{\link{plot_spct_file}},
  \code{\link{read_files2mspct}},
  \code{\link{read_oo_data}}, \code{\link{read_oo_ovdata}},
  \code{\link{read_oo_pidata}},
  \code{\link{read_oo_ssdata}},
  \code{\link{set_oo_ssdata_descriptor}},
  \code{\link{set_oo_ssdata_settings}}
}
\concept{functions for importing spectral data from files}
