% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file-utils.R
\name{plot_spct_file}
\alias{plot_spct_file}
\title{Read a file and plot spectrum}
\usage{
plot_spct_file(file, range = NULL, strength = 0)
}
\arguments{
\item{file}{a (readable binary-mode) connection or a character string giving
the name of the file to load (when tilde expansion is done).}

\item{range}{a numeric vector of length two, or any other object for which
function range() will return two.}

\item{strength}{numeric value to adjust the degree of smoothing.}
}
\description{
Read an .Rda file as saved during data acquisition and plot it, possibly
first trimming the range of wavelengths and/or smoothing the spectral data.
}
\seealso{
Function \code{\link[photobiology]{trim_wl}} is used to trim the
  range of the data to plot, and function
  \code{\link[photobiology]{smooth_spct}} with \code{method = "supsmu"} are
  used in the implementation, and prameters \code{range} and \code{strength}
  are passed to them, respectively.
}
