% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-continuous-interactive.R
\name{protocol_interactive}
\alias{protocol_interactive}
\title{Interactively select a measurement protocol}
\usage{
protocol_interactive(protocols)
}
\description{
Choose a protocol by name from a list of protocols, allowing the user to
correct the selection if needed.
}
\keyword{internal}
