% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2corr-cps.R
\name{raw2corr_cps}
\alias{raw2corr_cps}
\alias{raw2corr_cps.default}
\alias{raw2corr_cps.raw_spct}
\alias{raw2corr_cps.raw_mspct}
\title{Convert raw detector counts into counts-per-second}
\usage{
raw2corr_cps(x, ref.pixs.range, ...)

\method{raw2corr_cps}{default}(x, ref.pixs.range = NULL, ...)

\method{raw2corr_cps}{raw_spct}(
  x,
  ref.pixs.range = c(1, 100),
  despike = FALSE,
  hdr.tolerance = getOption("ooacquire.hdr.tolerance", default = 0.05),
  ...
)

\method{raw2corr_cps}{raw_mspct}(x, ref.pixs.range = c(1, 100), despike = FALSE, ...)
}
\arguments{
\item{x}{raw_spct object.}

\item{ref.pixs.range}{integer vector of length 2.}

\item{...}{passed to \code{photobiology::despike}.}

\item{despike}{logical flag, if TRUE despiking will be attempted.}

\item{hdr.tolerance}{numeric Passed as tolerance argument to merge_cps().}
}
\value{
a cps_spct object with one spectrum preserving the metadata present in
  \code{x}.
}
\description{
Replace data from bad pixels with interpolated values, replace data from
saturated and nearby pixels with NAs, apply linearization function if data
is not already linearized, optionally use a range of pixels as dark
reference, convert the raw counts for each integration time used into
counts-per-second, if data from bracketed integration times is available,
splice the different spectra.
}
\section{Methods (by class)}{
\itemize{
\item \code{raw2corr_cps(default)}: Default method

\item \code{raw2corr_cps(raw_spct)}: raw_spct method

\item \code{raw2corr_cps(raw_mspct)}: raw_spct method

}}
\seealso{
\code{\link[photobiology]{despike}()}

Other functions for conversion of raw-counts data:
\code{\link[=raw2cps]{raw2cps()}},
\code{\link[=s_fraction_corrected]{s_fraction_corrected()}},
\code{\link[=s_irrad_corrected]{s_irrad_corrected()}},
\code{\link[=s_irrad_update]{s_irrad_update()}}
}
\concept{functions for conversion of raw-counts data}
