% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2corr-cps.R
\name{raw2corr_cps}
\alias{raw2corr_cps}
\title{Convert raw detector counts into counts-per-second and correct}
\usage{
raw2corr_cps(x, ref.pixs.range = c(1, 100))
}
\arguments{
\item{x}{raw_spct object.}

\item{ref.pixs.range}{integer vector of length 2.}
}
\value{
a cps_spct object with one spectrum preserving the metadata present in
  \code{x}.
}
\description{
Replace data from bad pixels with interpolated values, replace data from
saturated and nearby pixels withs NAs, apply linearization function if data
is not already linearized, optionally use a range of pixels as dark
reference, convert the raw counts for each integration time used into
counts-per-second, if data from bracketed intergartion times is available,
splice the different spectra.
}

