% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2cps.R
\name{raw2cps}
\alias{raw2cps}
\alias{raw2cps.default}
\alias{raw2cps.raw_spct}
\alias{raw2cps.raw_mspct}
\title{Convert raw detector counts into counts per second (cps).}
\usage{
raw2cps(x, ...)

\method{raw2cps}{default}(x, ...)

\method{raw2cps}{raw_spct}(x, ...)

\method{raw2cps}{raw_mspct}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{not used in current version}
}
\value{
an object of class "cps_spct"

an object of class "cps_mspct"
}
\description{
These functions simply divide "raw counts" by the integration time used
for acquisition.
}
\section{Methods (by class)}{
\itemize{
\item \code{raw2cps(default)}: Default method

\item \code{raw2cps(raw_spct)}: Method for spectral data expressed as raw instrument counts.

\item \code{raw2cps(raw_mspct)}: Method for collections of raw-counts spectra

}}
\note{
In the case of objects of class "raw_spct" the columns with names
starting with "counts" are processed and renamed to columns with names
starting with "cps". All other columns are left unchanded.
}
\seealso{
Other functions for conversion of raw-counts data:
\code{\link[=raw2corr_cps]{raw2corr_cps()}},
\code{\link[=s_fraction_corrected]{s_fraction_corrected()}},
\code{\link[=s_irrad_corrected]{s_irrad_corrected()}},
\code{\link[=s_irrad_corrected_updt]{s_irrad_corrected_updt()}}
}
\concept{functions for conversion of raw-counts data}
