% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw2cps.R
\name{raw2cps}
\alias{raw2cps}
\alias{raw2cps.default}
\alias{raw2cps.numeric}
\alias{raw2cps.raw_spct}
\alias{raw2cps.raw_mspct}
\title{Convert raw detector counts into counts per second (cps).}
\usage{
raw2cps(x, ...)

\method{raw2cps}{default}(x, ...)

\method{raw2cps}{numeric}(x, integ.time, ...)

\method{raw2cps}{raw_spct}(x, ...)

\method{raw2cps}{raw_mspct}(x, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{not used in current version}

\item{integ.time}{numeric, the integration time used expressed in seconds.}
}
\value{
a numeric vector of the same length as x

an object of class "cps_spct"

an object of class "cps_mspct"
}
\description{
These functions simply divide "raw counts" by the integration time used
for acquisition.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method

\item \code{numeric}: Numeric method

\item \code{raw_spct}: Method for spectral data expressed as raw instrument counts.

\item \code{raw_mspct}: Method for collections of raw-counts spectra
}}

\note{
In the case of objects of class "raw_spct" the columns with names
starting with "counts" are processed and renamed to columns with names
starting with "cps". All other columns are left unchanded.
}
