% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-raw-data.R
\docType{data}
\name{red_filter.raw_mspct}
\alias{red_filter.raw_mspct}
\title{Raw counts data for a filter measurement}
\format{A raw_mscpt}
\usage{
red_filter.raw_mspct
}
\description{
A red long-pass filter from Heliopan (695 nm) measured using a household
incandescent lamp as light source. The spectrometer used was an Ocean Optics
Maya2000 Pro.
}
\keyword{datasets}
