% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref-correction.R
\name{ref_correction}
\alias{ref_correction}
\alias{ref_correction.cps_spct}
\alias{ref_correction.default}
\alias{ref_correction.numeric}
\alias{ref_correction.raw_spct}
\title{Convert raw detector counts into counts per second (cps).}
\usage{
ref_correction(x, y, .oper, ...)

\method{ref_correction}{default}(x, y, .oper, ...)

\method{ref_correction}{numeric}(x, y, .oper = `-`, ...)

\method{ref_correction}{raw_spct}(x, y, .oper = `-`, ...)

\method{ref_correction}{cps_spct}(x, y, .oper = `-`, ...)
}
\arguments{
\item{x, y}{R objects}

\item{.oper}{a function with its first two formal parameters accepting
numerical vectors of equal length (e.g. a binary numerical operator).}

\item{...}{not used in current version}
}
\value{
a numeric vector of the same length as x

an object of class "cps_spct"

an object of class "cps_spct"
}
\description{
These functions simply divide "raw counts" by the integration time used
for acquisition.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method

\item \code{numeric}: Numeric method

\item \code{raw_spct}: Method for spectral data expressed as raw
instrument counts.

\item \code{cps_spct}: Method for spectral data expressed as
counts per second.
}}
\note{
In the case of objects of class "raw_spct" the columns with names
starting with "counts" are processed. All other columns are left unchanded.
}

