% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-instr-desc.R
\name{rm_jwrapper}
\alias{rm_jwrapper}
\title{Remove java wrapper from descriptor}
\usage{
rm_jwrapper(x)
}
\arguments{
\item{x}{generic_spct or generic_mspct object with attribute
\code{instr.desc} set.}
}
\value{
a copy of \code{x} possibly with field \code{"w"} removed, if present,
  from the embedded \code{instr.desc} attribute of spectral objects.
}
\description{
Clean up left-behind wrappers used to communicate with Java code using
'rJava'.
}
\details{
Field \code{w}  contains a wrapper on a Java object used to communicate with
the OmniDriver driver during data acquisition. Once the current connection
between R and a spectrometer ends, these wrappers are invalidated, becoming
useless. However, if present they create a dependency on 'rJava', possibly
triggering errors. In recent versions of 'ooacquire' this wrapper is removed
immediately after acquisition. However, the instrument descriptor of spectral
objects created with versions of 'ooacquire' from some years ago can contain a
member storing the useless Java wrapper. This function removes this field if
present.
}
\note{
Method \code{getInstrDesc()} removes member field \code{w} from the returned
value but does not modify its argument.
}
