% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update-instr-desc.R
\name{rm_jwrapper}
\alias{rm_jwrapper}
\title{Remove java wrapper from descriptor}
\usage{
rm_jwrapper(x, verbose = getOption("photobiology.verbose", default = FALSE))
}
\arguments{
\item{x}{generic_spct or generic_mspct object with attribute
\code{instr.desc} set.}

\item{verbose}{logical If \code{TRUE}, issue message when no action is
needed, i.e., the spectral object does not contain a Java wrapper.}
}
\value{
a copy of \code{x} possibly with field \code{"w"} removed, if present,
  from the embedded \code{instr.desc} attribute of spectral objects.
}
\description{
Clean up left-behind wrappers used to communicate with Java code using
'rJava'.
}
\details{
Field \code{w}  contains a wrapper on a Java object used to communicate with
the OmniDriver driver during data acquisition. Once the current connection
between R and a spectrometer ends, these wrappers are invalidated, becoming
useless. However, if present they create a dependency on 'rJava', possibly
triggering errors. In recent versions of 'ooacquire' this wrapper is removed
immediately after acquisition. However, the instrument descriptor of spectral
objects created with versions of 'ooacquire' from some years ago can contain a
member storing the useless Java wrapper. This function removes this field if
present.

Method \code{getInstrDesc()} removes member field \code{w} from the returned
value but does not modify its argument, it returns the modified object.
}
\seealso{
Use function \code{\link{files_rm_jwrapper}()} to apply function
  \code{rm_jwrapper()} to objects stored in .rda files,

Other Functions and methods related to instrument descriptors.:
\code{\link[=files_rm_jwrapper]{files_rm_jwrapper()}},
\code{\link[=files_s_irrad_update]{files_s_irrad_update()}},
\code{\link[=instr_defaults]{instr_defaults()}},
\code{\link[=update_bad_pixs]{update_bad_pixs()}}
}
\concept{Functions and methods related to instrument descriptors.}
