% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s-fraction-corrected.R
\name{s_fraction_corrected}
\alias{s_fraction_corrected}
\alias{s_fraction_corrected.default}
\alias{s_fraction_corrected.list}
\alias{s_fraction_corrected.raw_mspct}
\title{Convert raw counts data into a spectral fraction}
\usage{
s_fraction_corrected(x, ...)

\method{s_fraction_corrected}{default}(x, ...)

\method{s_fraction_corrected}{list}(
  x,
  reference.value = 1,
  type = "internal",
  time = NULL,
  correction.method,
  qty.out = "Tfr",
  descriptor = NULL,
  dyn.range = NULL,
  locale = NULL,
  verbose = getOption("photobiology.verbose", default = FALSE),
  ...
)

\method{s_fraction_corrected}{raw_mspct}(
  x,
  spct.names = c(sample = "sample", reference = "reference", dark = "dark"),
  reference.value = 1,
  type = switch(qty.out, Tfr = "internal", Rfr = "total"),
  correction.method,
  dyn.range = NULL,
  qty.out = "Tfr",
  verbose = getOption("photobiology.verbose", default = FALSE),
  ...
)
}
\arguments{
\item{x}{A named list of one to three vectors of file names, with names
"sample", "reference", and "dark". Or a raw_mspt object, or a raw_spct object.}

\item{...}{Named argument passed to \code{photobiology::cps2irrad} which is
the final calculation step.}

\item{reference.value}{numeric or filter_spct or reflector_spct object, with the
fractional transmittance or reflectance of the reference.}

\item{type}{character One of "internal" or "total".}

\item{time}{a \code{POSIXct} object, but if \code{NULL} the date stored in
file is used, and if \code{NA} no date variable is added}

\item{correction.method}{A named list of constants and functions defining the
method to be sued for stray light and dark signal corrections.}

\item{qty.out}{character, one of "Tfr", "Rfr".}

\item{descriptor}{A named list with a descriptor of the characteristics of
the spectrometer (if serial number does not agree an error is triggered).}

\item{dyn.range}{numeric Effective dynamic range of the spectrometer.}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{verbose}{Logical indicating the level of warnings wanted.}

\item{spct.names}{named character vector of length three, to map names in
\code{x} to those expected.}
}
\description{
Convert raw counts data into spectral transmittance or spectral reflectance.
}
\section{Methods (by class)}{
\itemize{
\item \code{s_fraction_corrected(default)}: Default for generic function.

\item \code{s_fraction_corrected(list)}: Default for generic function.

\item \code{s_fraction_corrected(raw_mspct)}: Default for generic function.

}}
\note{
Currently \code{s_fraction_corrected.list} allows processing of files
  written by OceanOptics' SpectraSuite software, from protocols with
  integration-time bracketing or not, with a dark reference measurement or
  not. Four measurements components are recognized: a "sample" measurement,
  a "referenece" measurement using a clear or white, a "filter" measurement
  with a UV-blocking filter in the light pass, and a "dark" measurement.
  Only "sample" and "reference" are mandatory. Data should be raw counts,
  either corrected for detector non-linearity or not. All three spectra
  should be acquired using the same instrument settings to achieve good
  accuracy.
}
\seealso{
Other functions for conversion of raw-counts data:
\code{\link[=raw2corr_cps]{raw2corr_cps()}},
\code{\link[=raw2cps]{raw2cps()}},
\code{\link[=s_irrad_corrected]{s_irrad_corrected()}},
\code{\link[=s_irrad_update]{s_irrad_update()}}
}
\concept{functions for conversion of raw-counts data}
