% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_trans_corrected.R
\name{s_fraction_corrected}
\alias{s_fraction_corrected}
\alias{s_fraction_corrected.default}
\alias{s_fraction_corrected.list}
\alias{s_fraction_corrected.raw_mspct}
\title{Convert raw counts data into a spectral fraction}
\usage{
s_fraction_corrected(x, ...)

\method{s_fraction_corrected}{default}(x, ...)

\method{s_fraction_corrected}{list}(x, ref.value = 1, ref.type = "internal",
  time = NULL, method, qty.out = "Tfr", descriptor = NULL,
  locale = NULL, verbose = FALSE, ...)

\method{s_fraction_corrected}{raw_mspct}(x, ref.value = 1,
  ref.type = "internal", method, qty.out = "Tfr", verbose = FALSE, ...)
}
\arguments{
\item{x}{A named list of one to three vectors of file names, with names
"sample", "ref", and "dark".}

\item{...}{Named argument passed to \code{photobiology::cps2irrad} which is
the final calculation step.}

\item{ref.value}{numeric the fractional transmittance or reflectance of the
reference.}

\item{ref.type}{character One of "internal" or "total".}

\item{time}{a \code{POSIXct} object, but if \code{NULL} the date stored in
file is used, and if \code{NA} no date variable is added}

\item{method}{A named list of constants and functions defining the
method to be sued for stray light and dark signal corrections.}

\item{qty.out}{character, one of "Tfr", "Rfr".}

\item{descriptor}{A named list with a descriptor of the characteristics of
the spectrometer (if serial number does not agree an error is triggered).}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{verbose}{Logical indicating the level of warnings wanted.}
}
\description{
Convert raw counts data into spectral transmittance or spectral reflectance.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default for generic function.

\item \code{list}: Default for generic function.

\item \code{raw_mspct}: Default for generic function.
}}
\note{
Currently \code{s_fraction_corrected.list} allows processing of files
  written by OceanOptics' SpectraSuite software, from protocols with
  integration-time bracketing or not, with a dark reference measurement or
  not. Four measurements components are recognized: a "sample" measurement,
  a "referenece" measurement using a clear or white, a "filter" measurement
  with a UV-blocking filter in the light pass, and a "dark" measurement.
  Only "sample" and "reference" are mandatory. Data should be raw counts,
  either corrected for detector non-linearity or not. All three spectra
  should be acquired using the same instrument settings to achieve good
  accuracy.
}

