% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s-irrad-corrected.R
\name{s_irrad_corrected}
\alias{s_irrad_corrected}
\alias{s_irrad_corrected.default}
\alias{s_irrad_corrected.list}
\alias{s_irrad_corrected.raw_mspct}
\alias{s_irrad_corrected.raw_spct}
\title{Convert raw counts data into spectral irradiance or fluence}
\usage{
s_irrad_corrected(x, ...)

\method{s_irrad_corrected}{default}(x, ...)

\method{s_irrad_corrected}{list}(
  x,
  time = NULL,
  correction.method,
  return.cps = FALSE,
  descriptor,
  locale = NULL,
  verbose = getOption("photobiology.verbose", default = FALSE),
  ...
)

\method{s_irrad_corrected}{raw_mspct}(
  x,
  spct.names = c(light = "light", filter = "filter", dark = "dark"),
  correction.method,
  return.cps = FALSE,
  verbose = getOption("photobiology.verbose", default = FALSE),
  ...
)

\method{s_irrad_corrected}{raw_spct}(
  x,
  time = NULL,
  correction.method,
  return.cps = FALSE,
  verbose = getOption("photobiology.verbose", default = FALSE),
  ...
)
}
\arguments{
\item{x}{A named list of one to three vectors of file names, with names
"light", "filter", and "dark". Or a raw_mspt object, or a raw_spct object.}

\item{...}{Named argument passed to \code{photobiology::cps2irrad} which is
the final calculation step.}

\item{time}{a \code{POSIXct} object, but if \code{NULL} the date stored in
file is used, and if \code{NA}, date is set to NA.}

\item{correction.method}{A named list of constants and functions defining the
method to be used for stray light and dark signal corrections.}

\item{return.cps}{logical Useful when there is no need to apply a calibration,
such as when computing new calibration multipliers.}

\item{descriptor}{A named list with a descriptor of the characteristics of
the spectrometer (if serial number does not agree an error is triggered).}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{verbose}{Logical indicating the level of warnings wanted.}

\item{spct.names}{named character vector of length three, to map names in
\code{x} to those expected.}
}
\description{
Convert raw counts data into spectral irradiance or fluence
}
\section{Methods (by class)}{
\itemize{
\item \code{s_irrad_corrected(default)}: Default for generic function.

\item \code{s_irrad_corrected(list)}: Default for generic function.

\item \code{s_irrad_corrected(raw_mspct)}: Default for generic function.

\item \code{s_irrad_corrected(raw_spct)}: Default for generic function.

}}
\note{
Currently \code{s_irrad_corrected.list} allows processing of files
  written by OceanOptics' SpectraSuite software, from protocols with
  integration-time bracketing or not, with a dark reference measurement or
  not. Three measurements components are recognized: a "light" measurement, a
  "filter" measurement using a polycarbonate filter and a dark measurement.
  Only "light" is mandatory. Data should be raw counts, either corrected for
  detector non-linearity or not. All three spectra should be acquired using
  the same instrument settings to achieve good accuracy.
}
\seealso{
Other functions for conversion of raw-counts data: 
\code{\link{raw2corr_cps}()},
\code{\link{raw2cps}()},
\code{\link{s_fraction_corrected}()}
}
\concept{functions for conversion of raw-counts data}
