% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb-corrections.R
\name{s_irrad_corrected}
\alias{s_irrad_corrected}
\alias{s_irrad_corrected.default}
\alias{s_irrad_corrected.list}
\alias{s_irrad_corrected.raw_mspct}
\title{Convert raw counts data into spectral irradiance}
\usage{
s_irrad_corrected(x, ...)

\method{s_irrad_corrected}{default}(x, ...)

\method{s_irrad_corrected}{list}(x, method.data, descriptor, locale,
  verbose = FALSE, ...)

\method{s_irrad_corrected}{raw_mspct}(x, method.data, verbose = FALSE, ...)
}
\arguments{
\item{x}{A named list of one to hree vectors of file names, with names
"light", "filter", and "dark".}

\item{method.data}{A named list of constants and functions defining the
method to be sued for stray light and dark signal corrections.}

\item{descriptor}{A named list with a descriptor of the characteristics of
the spectrometer (if serial number does not agree an error is triggered).}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}

\item{verbose}{Logical indicating the level of warnings wanted.}
}
\description{
Convert raw counts data into spectral irradiance
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default for generic function.

\item \code{list}: Default for generic function.

\item \code{raw_mspct}: Default for generic function.
}}
\note{
Currently \code{s_irrad_corrected.list} allows processing of files
  written by OceanOptics' SpectraSuite software, from protocols with
  integration-time bracketing or not, with a dark reference measurement or
  not. Three measurements components are recognized: a "light" measurement, a
  "filter" measurement using a polycarbonate filter and a dark measurement.
  Only "light" is mandatory. Data should be raw counts, either corrected for
  detector non-linearity or not. All three spectra should be acquired using
  the same instrument settings to achieve good accuracy.
}

