% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{set_attributes_interactive}
\alias{set_attributes_interactive}
\title{Interactively set user attributes}
\usage{
set_attributes_interactive(user.attrs = list(what.measured = "",
  comment.text = ""))
}
\arguments{
\item{user.attrs}{character Default values for the attributes.}
}
\description{
Allow user to provide values for "user supplied" attribute values
\code{"comment"} and \code{"what.measured"}.
}
