% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-continuous-interactive.R
\name{set_attributes_interactive}
\alias{set_attributes_interactive}
\title{Interactively set user attributes}
\usage{
set_attributes_interactive(user.attrs)
}
\description{
Enter values for "user supplied" attribute values.
}
\keyword{internal}
