% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_bad_pixs}
\alias{set_descriptor_bad_pixs}
\title{Add bad pixel informatoin to an instrument description}
\usage{
set_descriptor_bad_pixs(descriptor, bad.pixs)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{bad.pixs}{numeric vector of sorted wavelengths values corresponding to each
pixel in the instrument array.}
}
\value{
a copy of the argument passed for \code{oo_descriptor} with the
wavelengths field of the calibration data replaced by the new values.
}
\description{
A integer vector of indexes to bad pixels in the instrument array. Data from
these array pixels will be discarded.
}

