% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_integ_time}
\alias{set_descriptor_integ_time}
\title{Replace integration time limits in instrument descriptor}
\usage{
set_descriptor_integ_time(
  descriptor,
  min.integ.time = NA_integer_,
  max.integ.time = NA_integer_,
  force.change = FALSE
)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{min.integ.time, max.integ.time}{numeric values in seconds in seconds.}

\item{force.change}{If FALSE values outside the range returned by a query
to the instrument trigger an error. If TRUE this test is overriden.}
}
\value{
a copy of the argument passed for \code{oo_descriptor} with the
integration time fields of the descriptor modified.
}
\description{
This function can be needed in exceptional cases such as when the limits
stored in the intrument's persistent memory are wrong. In other cases it can
be used to restrict the range of values allowed to be set to a smaller range
than natively supported by the spectrometer.
}
\note{
This function should not be needed, but for some instruments the query
may fail or return the wrong value. Values should be within the range
in the instrument's specifications. Setting wrong values can result in
invalid data without an error being triggered.
}
\seealso{
Other Functions and methods related to instrument descriptors:
\code{\link[=get_oo_descriptor]{get_oo_descriptor()}},
\code{\link[=set_descriptor_bad_pixs]{set_descriptor_bad_pixs()}},
\code{\link[=set_descriptor_entrance_optics]{set_descriptor_entrance_optics()}},
\code{\link[=set_descriptor_irrad_mult]{set_descriptor_irrad_mult()}},
\code{\link[=set_descriptor_nl]{set_descriptor_nl()}},
\code{\link[=set_descriptor_wl]{set_descriptor_wl()}}
}
\concept{Functions and methods related to instrument descriptors}
