% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_irrad_mult}
\alias{set_descriptor_irrad_mult}
\title{Add spectral irradiance calibration}
\usage{
set_descriptor_irrad_mult(
  descriptor,
  irrad.mult,
  wl.range = NULL,
  start.date = lubridate::today(tzone = "UTC") - lubridate::days(1),
  end.date = lubridate::today(tzone = "UTC") + lubridate::days(1),
  cal.source = "unknown"
)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{irrad.mult}{numeric vector of the same length as the number of pixels
or of length one.}

\item{wl.range}{numeric Range of wavelengths for which the calibration is
valid.}

\item{start.date, end.date}{range of dates when calibration is valid.}

\item{cal.source}{character Identifier or source of the calibration data,
used for traceability.}
}
\value{
A copy of the argument passed for \code{oo_descriptor} with the
  irrad.mult field of the calibration data replaced by the new values.
}
\description{
Adds calibration data expressed as multipliers for each pixel stored in a
numeric vector.
}
\seealso{
Other Functions and methods related to instrument descriptors:
\code{\link[=get_oo_descriptor]{get_oo_descriptor()}},
\code{\link[=set_descriptor_bad_pixs]{set_descriptor_bad_pixs()}},
\code{\link[=set_descriptor_entrance_optics]{set_descriptor_entrance_optics()}},
\code{\link[=set_descriptor_integ_time]{set_descriptor_integ_time()}},
\code{\link[=set_descriptor_nl]{set_descriptor_nl()}},
\code{\link[=set_descriptor_wl]{set_descriptor_wl()}}
}
\concept{Functions and methods related to instrument descriptors}
