% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_irrad_mult}
\alias{set_descriptor_irrad_mult}
\title{Add spectral irradiance calibration}
\usage{
set_descriptor_irrad_mult(oo_descriptor, irrad.mult, wl.range = NULL,
  start.date = lubridate::today() - lubridate::days(1),
  end.date = lubridate::today() + lubridate::days(1))
}
\arguments{
\item{oo_descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{irrad.mult}{numeric vector of the same length as the number of pixels
or of length one.}

\item{wl.range}{numeric Range of wavelengths for which the calibration is
valid.}

\item{start.date, end.date}{range of dates when calibration is valid.}
}
\value{
A copy of the argument passed for \code{oo_descriptor} with the
irrad.mult field of the calibration data replaced by the new values.
}
\description{
Adds calibration data for each pixel as a numeric vector
}

