% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_nl}
\alias{set_descriptor_nl}
\title{Replace linearization function in instrument description.}
\usage{
set_descriptor_nl(descriptor, nl.coeff = NA_real_, nl.fun = NULL)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{nl.coeff}{numeric vector, if nl.fun is missing, assumed to be a
polynomial.}

\item{nl.fun}{A function or a polynom::polynomial object containing
the linearization to be applied.}
}
\value{
A copy of the argument passed for \code{oo_descriptor} with the
wavelengths field of the calibration data replaced by the new values.
}
\description{
Uses a user supplied function, possibly that supplied by a manufacturer
like Ocean Optics stored in firmware or in any other form.
}
\seealso{
Other Functions and methods for instrument descriptors:
\code{\link[=get_oo_descriptor]{get_oo_descriptor()}},
\code{\link[=set_descriptor_bad_pixs]{set_descriptor_bad_pixs()}},
\code{\link[=set_descriptor_entrance_optics]{set_descriptor_entrance_optics()}},
\code{\link[=set_descriptor_integ_time]{set_descriptor_integ_time()}},
\code{\link[=set_descriptor_irrad_mult]{set_descriptor_irrad_mult()}},
\code{\link[=set_descriptor_wl]{set_descriptor_wl()}}
}
\concept{Functions and methods for instrument descriptors}
