% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_nl}
\alias{set_descriptor_nl}
\title{Replace linearization function in instrument description.}
\usage{
set_descriptor_nl(descriptor, nl.fun)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{nl.fun}{A function or a polynom::polynomial object containing
the linearization to be applied.}
}
\value{
A copy of the argument passed for \code{oo_descriptor} with the
wavelengths field of the calibration data replaced by the new values.
}
\description{
Uses a user supplied function, possibly that supplied by a manufacturer
like Ocean Optics stored in firmware or in any other form.
}
