% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oo-instrument.R
\name{set_descriptor_wl}
\alias{set_descriptor_wl}
\title{Replace wavelength values in an instrument description}
\usage{
set_descriptor_wl(descriptor, wl)
}
\arguments{
\item{descriptor}{list as returned by function \code{get_oo_descriptor}}

\item{wl}{numeric vector of sorted wavelengths values corresponding to each
pixel in the instrument array.}
}
\value{
a copy of the argument passed for \code{oo_descriptor} with the
wavelengths field of the calibration data replaced by the new values.
}
\description{
Replace wavelength values in an instrument descriptor for an Ocean Optics
spectrometer with new values if valid. (e.g. when wavelngth calibration
is not stored in firmware).
}

