% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{set_folder_interactive}
\alias{set_folder_interactive}
\title{Interactively get folder to use}
\usage{
set_folder_interactive(folder.name = ".")
}
\arguments{
\item{folder.name}{character Default name of the folder.}
}
\description{
Enter values for "user supplied" folder.
}
\details{
If the requested folder does not already exist it will be created.
The name of the folder is returned, but NOT set as working directory.
}
\seealso{
Other interactive acquisition utility functions: \code{\link{choose_ch_interactive}},
  \code{\link{choose_sr_interactive}},
  \code{\link{f.trigger.message}},
  \code{\link{list_srs_interactive}},
  \code{\link{protocol_interactive}},
  \code{\link{set_attributes_interactive}},
  \code{\link{set_seq_interactive}},
  \code{\link{tune_interactive}}
}
\concept{interactive acquisition utility functions}
