% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive.R
\name{set_folder_interactive}
\alias{set_folder_interactive}
\title{Interactively get folder to use}
\usage{
set_folder_interactive(folder.name = ".")
}
\description{
Enter values for "user supplied" folder.
}
\keyword{internal}
