% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-settings.R
\name{set_integ_time}
\alias{set_integ_time}
\title{Set integration time}
\usage{
set_integ_time(acq.settings, integ.time, verbose = TRUE)
}
\arguments{
\item{acq.settings}{list as returned by function \code{acq_settings()}}
}
\value{
a copy of the argument passed for \code{acq.settings} with the
\code{integ.time} field of the settings data replaced by \code{integ.time}.
}
\description{
Tag the settings as linearized
}
\keyword{internal}
