% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-settings.R
\name{set_linearized}
\alias{set_linearized}
\title{Set linearized attribute}
\usage{
set_linearized(acq.settings)
}
\arguments{
\item{acq.settings}{list as returned by function \code{acq_settings()}}
}
\value{
a copy of the argument passed for \code{acq.settings} with the
linearized field of the settings data replaced by TRUE.
}
\description{
Tag the settings as linearized
}
\keyword{internal}

