% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-settings.R
\name{set_num_exposures}
\alias{set_num_exposures}
\title{Set number of exposures}
\usage{
set_num_exposures(acq.settings, num.exposures = NA_integer_,
  single.scan = FALSE, verbose = TRUE)
}
\arguments{
\item{acq.settings}{list as returned by function \code{acq_settings()}}
}
\value{
a copy of the argument passed for \code{acq.settings} with the
\code{integ.time} field of the settings data replaced by \code{integ.time}.
}
\description{
Set the number of exposures in the instrument settings data structure. This
value is by default missing, indicating a protocol in which one assumes
contant illumination during an integration. If the light exposure is shorter
than the integration time, we need to express the results per exposure, and
consequently calculations done based on the number of exposure during each
integration. This is the case, for example, when measuring the spectral
emission of a xenon flash.
}
\keyword{internal}
