% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-oo-header.R
\name{set_oo_ssdata_descriptor}
\alias{set_oo_ssdata_descriptor}
\title{Set the instrument description.}
\usage{
set_oo_ssdata_descriptor(x, file.header = comment(x), descriptor = NULL,
  action = "overwrite")
}
\arguments{
\item{x}{generic_spct, although with defaults only raw_spct.}

\item{file.header}{character string The header of the file output by
SpectraSuite.}

\item{descriptor}{list An already built instrument descriptor,
in which case \code{file.header} is used only to validate the serial
number.}

\item{action}{character A flag indicating if an existing instrument
descriptor in \code{x} should be overwritten, merged or kept as is.}
}
\value{
A copy of \code{x} with updated attributes.
}
\description{
Model and serial number are retrieved from the file header, the remaining
unknown parameter values are set to NA. These values are used to set
the "inst.desc" atrribute of \code{x}.
}

