% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-oo-header.R
\name{set_oo_ssdata_settings}
\alias{set_oo_ssdata_settings}
\title{Set the values of instrument settings from file header}
\usage{
set_oo_ssdata_settings(x, file.header = comment(x), overwrite = TRUE)
}
\arguments{
\item{x}{generic_spct, although with defaults only raw_spct.}

\item{file.header}{character string The header of the file output by
SpectraSuite.}

\item{overwrite}{logical A flag indicating if an existing valid instrument
descriptor in \code{x} should be overwritten.}
}
\value{
A copy of \code{x} with updated attributes.
}
\description{
Parse the header of the file returned by SpectraSuite for corrections,
smotthing and acquisition parameters used. These values are used to set
the "inst.settings" atrribute of \code{x}.
}

