% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{set_seq_interactive}
\alias{set_seq_interactive}
\title{Interactively set sequential measurements}
\usage{
set_seq_interactive(
  seq.settings = list(start.boundary = "second", initial.delay = 0, step.delay = 0,
    num.steps = 1),
  measurement.duration = 0,
  minimum.step.delay = measurement.duration,
  time.division = 0
)
}
\arguments{
\item{seq.settings}{numeric Definition of time steps for a sequence of
repeated measurements. Named vector with member named
\code{start.boundary}, \code{"initial.delay"}, \code{"step.delay"}, and
\code{"num.steps"}.}

\item{measurement.duration}{numeric Duration of one measurement event (s).}

\item{minimum.step.delay}{numeric Minimum duration of \code{"step.delay"} (s).}

\item{time.division}{numeric The step is forced to be a multiple of this
time duration, because spectrometers normally are constantly acquiring
spectra and thbey return the most recently acquired one. Should be set to
the integration time plus a very small overhead (s).}
}
\value{
A named numeric vector of length two.
}
\description{
Enter settings defining a sequence of spectra to be measured as a time
series.
}
\details{
Function \code{seq.settings()} allows users to enter values needed
to define a sequence of spectral acquisitions. These are the time unit
boundary to synchronize to, the delay or duration of the time
step between successive acquisitions and the number of acquisitions in the
series. The \code{measurement.time} determines the minimum length for
\code{"step"}.

A sequence of measurements are expected to share a single reference or
dark scan, and be done as a sequence. With a single time point, the initial
delay or time unit boundary can be used to schedule a single timed
measurement.
}
\seealso{
Other interactive acquisition utility functions:
\code{\link[=choose_ch_interactive]{choose_ch_interactive()}},
\code{\link[=choose_sr_interactive]{choose_sr_interactive()}},
\code{\link[=f.trigger.message]{f.trigger.message()}},
\code{\link[=list_srs_interactive]{list_srs_interactive()}},
\code{\link[=protocol_interactive]{protocol_interactive()}},
\code{\link[=set_attributes_interactive]{set_attributes_interactive()}},
\code{\link[=set_folder_interactive]{set_folder_interactive()}},
\code{\link[=set_session_name_interactive]{set_session_name_interactive()}},
\code{\link[=set_user_name_interactive]{set_user_name_interactive()}},
\code{\link[=set_wlrange_interactive]{set_wlrange_interactive()}},
\code{\link[=tune_interactive]{tune_interactive()}}
}
\concept{interactive acquisition utility functions}
