% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{set_seq_interactive}
\alias{set_seq_interactive}
\title{Interactively set sequential measurements}
\usage{
set_seq_interactive(seq.settings = c(step.delay = 0, num.steps = 1L))
}
\arguments{
\item{seq.settings}{numeric Definition of time steps for a sequence of repeated
measurements. Named vector with member names \code{"step"}, and
\code{"steps"}.}
}
\description{
Adjust integration time settings, allowing the user to repeat the tunning,
and to change some of the parameters used for tunning.
}
