% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive.R
\name{set_seq_interactive}
\alias{set_seq_interactive}
\title{Interactively set sequential measurements}
\usage{
set_seq_interactive(seq.settings)
}
\description{
Adjust integration time settings, allowing the user to repeat the tunning,
and to change some of the parameters used for tunning.
}
\keyword{internal}
