% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{set_user_name_interactive}
\alias{set_user_name_interactive}
\title{Interactively get user name to set}
\usage{
set_user_name_interactive(user.name = NULL)
}
\arguments{
\item{used.name}{character Default name of the folder.}
}
\description{
Enter values for "user supplied" name.
}
\details{
Validate user name, and allow user to change the default.
}
\seealso{
Other interactive acquisition utility functions: 
\code{\link{choose_ch_interactive}()},
\code{\link{choose_sr_interactive}()},
\code{\link{f.trigger.message}()},
\code{\link{list_srs_interactive}()},
\code{\link{protocol_interactive}()},
\code{\link{set_attributes_interactive}()},
\code{\link{set_folder_interactive}()},
\code{\link{set_seq_interactive}()},
\code{\link{set_session_name_interactive}()},
\code{\link{tune_interactive}()}
}
\concept{interactive acquisition utility functions}
