% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session.R
\name{start_session}
\alias{start_session}
\title{Connect to Maya spectrometer}
\usage{
start_session(error.action = stop)
}
\arguments{
\item{error.action}{function, usually one of \code{stop()}, \code{warning()}
or \code{message()}.}
}
\value{
On success a java wrapper which allows access to the driver with an open
connection to the instrument.
}
\description{
Open a connection to the first spectrometer found
and initialize an object with a reference to the
Java object returned by Omni Driver.
}
\seealso{
Other spectrometer-connection functions: \code{\link{end_session}},
  \code{\link{list_instruments}}
}
\concept{spectrometer-connection functions}
