% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open.close.R
\name{start_session}
\alias{start_session}
\title{Connect to Maya spectrometer}
\usage{
start_session()
}
\description{
Open a connection to the first spectrometer found
and initialize an object with a reference to the
Java object returned by Omni Driver.
}

