% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-raw-data.R
\docType{data}
\name{sun001.raw_mspct}
\alias{sun001.raw_mspct}
\title{Raw counts data for a lamp measurement.}
\format{
A raw_mspct
}
\usage{
sun001.raw_mspct
}
\description{
Solar radiation at ground level measured in Helsinki Finland. The
spectrometer used was an Ocean Optics Maya2000 Pro.
}
\seealso{
Other objects containing example raw-counts data: 
\code{\link{blue_filter.raw_mspct}},
\code{\link{halogen.raw_mspct}},
\code{\link{red_filter.raw_mspct}},
\code{\link{white_LED.raw_mspct}},
\code{\link{xenon_flash.raw_mspct}}
}
\concept{objects containing example raw-counts data}
\keyword{datasets}
