% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive-utils.R
\name{tune_interactive}
\alias{tune_interactive}
\title{Interactively adjust the integration time settings}
\usage{
tune_interactive(descriptor, acq.settings, start.int.time = 0.1,
  interface.mode = "auto")
}
\arguments{
\item{descriptor}{list Descriptor of the instrument, including wrapper to
Java object used to acces the instrument.}

\item{acq.settings}{list containing starting values for instrument settings.}

\item{start.int.time}{numeric Integration time to use as starting guess when
tuning the settings.}

\item{interface.mode}{character One of "simple", "auto", or "manual".}
}
\description{
Adjust integration time settings, allowing the user to repeat the tunning,
and to change some of the parameters used for tunning.
}
