% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acq-interactive.R
\name{tune_interactive}
\alias{tune_interactive}
\title{Interactively adjust the integration time settings}
\usage{
tune_interactive(descriptor, settings, start.int.time)
}
\description{
Adjust integration time settings, allowing the user to repeat the tunning,
and to change some of the parameters used for tunning.
}
\keyword{internal}
