% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb-corrections.R
\name{uvb_corrections}
\alias{uvb_corrections}
\alias{slit_function_correction}
\title{Apply filter-based stray-light correction}
\usage{
uvb_corrections(
  x,
  spct.names = c(light = "light", filter = "filter", dark = "dark"),
  stray.light.method = "original",
  stray.light.wl = c(218.5, 228.5),
  flt.dark.wl = c(193, 209.5),
  flt.ref.wl = c(360, 379.5),
  flt.Tfr = 0.9,
  inst.dark.pixs = 1:4,
  worker.fun = NULL,
  trim = 0.05,
  hdr.tolerance = getOption("ooacquire.hdr.tolerance", default = 0.05),
  verbose = getOption("photobiology.verbose", default = FALSE),
  ...
)

slit_function_correction(
  x,
  worker.fun = NULL,
  hdr.tolerance = getOption("ooacquire.hdr.tolerance", default = 0.05),
  verbose = getOption("photobiology.verbose", default = FALSE),
  ...
)
}
\arguments{
\item{x}{\code{raw_mspct} The raw counts from measurements.}

\item{spct.names}{named character vector of length three.}

\item{stray.light.method}{Method variant used, "original" (Ylianttila),
"simple", "full", "sun", "raw", "none".}

\item{stray.light.wl}{numeric vector of length 2 giving the range of
wavelengths to use for the final stray light correction.}

\item{flt.dark.wl, flt.ref.wl}{numeric vectors of length 2 giving the ranges
of wavelengths to use for the "dark" and "illuminated" regions of the
array in the filter correction.}

\item{flt.Tfr}{numeric fractional transmittance of the filter to the source
of stray light, used only for method "simple".}

\item{inst.dark.pixs}{numeric vector with indexes to array pixels that
are in full darkness by instrument design.}

\item{worker.fun}{function actually doing the correction on the w.lengths and
counts per second vectors, or the name of the function as a character string.}

\item{trim}{a numeric value to be used as argument for mean}

\item{hdr.tolerance}{numeric Tolerance for mean deviation among cps columns as
a fraction of one. Used in check of HDR consistency.}

\item{verbose}{Logical indicating the level of warnings wanted.}

\item{...}{additional parameters passed to worker.fun.}
}
\value{
A \code{cps_spct} object with the corrected count-per-second values
  in coulmn "cps".
}
\description{
Apply to a counts-per-second spectrum corrections based of a paired reading
obtained with a polycarbonate filter (long-pass with cut-in at 400 nm). This
is a bit more sophisticated than simple subtracting the filter reading from
the measurement as the effect of the filter itself on stray light is corrected
for.
}
\note{
\code{stray.light.method = "none"} is a valid argument only for
  function \code{uvb_corrections()}. The default \code{worker.fun} is just an
  example. Corrections are specific to each individual spectrometer unit (not
  just type or configuration) and code needs to be written for most
  individual use cases. The slit function tail correction requires the
  characterization of the shape of the slit function by measuring one or more
  laser beams at suitable wavelengths.
}
\references{
\url{http://www.r4photobiology.info}
}
\seealso{
Other spectral data-processing functions:
\code{\link[=MAYP112785_tail_correction]{MAYP112785_tail_correction()}},
\code{\link[=MAYP11278_tail_correction]{MAYP11278_tail_correction()}},
\code{\link[=check_sn_match]{check_sn_match()}},
\code{\link[=linearize_counts]{linearize_counts()}},
\code{\link[=merge_raw_mspct]{merge_raw_mspct()}},
\code{\link[=new_correction_method]{new_correction_method()}},
\code{\link[=ref_correction]{ref_correction()}},
\code{\link[=trim_counts]{trim_counts()}}
}
\author{
Algorithm for method "original" developed by Lasse Ylianttila. Other
  methods are modified from Ylianttila's method by Pedro J. Aphalo.
}
\concept{spectral data-processing functions}
