% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb-corrections.R
\name{uvb_corrections}
\alias{filter_correction}
\alias{slit_function_correction}
\alias{uvb_corrections}
\alias{uvb_corrections_nb_files}
\title{Apply filter-based stray-light correction}
\usage{
uvb_corrections(x, flt, dark, method = "original", stray.light.wl = c(218.5,
  228.5), flt.dark.wl = c(193, 209.5), flt.ref.wl = c(360, 379.5),
  worker_fun = ooacquire::maya_tail_correction, trim = 0, verbose = FALSE,
  ...)

slit_function_correction(x, worker_fun = ooacquire::maya_tail_correction, ...)

filter_correction(x, flt, method = "original", flt.dark.wl = c(193, 209.5),
  flt.ref.wl = c(360, 379.5), trim = 0, verbose = FALSE)

uvb_corrections_nb_files(files, descriptor = descriptor,
  locale = readr::default_locale(), method = "original",
  stray.light.wl = c(218.5, 228.5), flt.dark.wl = c(193, 209.5),
  flt.ref.wl = c(360, 379.5), worker_fun = ooacquire::maya_tail_correction,
  trim = 0, verbose = FALSE)
}
\arguments{
\item{x, flt, dark}{\code{cps_spct} objects.}

\item{method}{Method variant used "original" (Ylianttila), "full" (Aphalo),
"sun" (full with trimming).}

\item{stray.light.wl}{numeric vector of length 2 giving the range of
wavelengths to use for the final stray light correction.}

\item{flt.dark.wl, flt.ref.wl}{numeric vectors of length 2 giving the ranges
of wavelengths to use for the "dark" and "illuminated" regions of the
array in the filter correction.}

\item{worker_fun}{function actually doing the correction on the w.lengths and
counts per second vectors.}

\item{trim}{a numeric value to be used as argument for mean}

\item{verbose}{Logical indicating the level of warnings wanted. Defaults to
\code{FALSE}.}

\item{...}{additional params passed to worker_fun.}

\item{files}{A named list of three file names, with names "light", "filter",
and "dark".}

\item{descriptor}{A named list with a descriptor of the characteristics of
the spectrometer (if serial number does not agree an error is triggered).}

\item{locale}{The locale controls defaults that vary from place to place. The
default locale is US-centric (like R), but you can use
\code{\link[readr]{locale}} to create your own locale that controls things
like the default time zone, encoding, decimal mark, big mark, and day/month
names.}
}
\value{
A copy of \code{x} with the "cps" data replaced by the corrected
  ones.
}
\description{
Apply to a counts-per-second spectrum corrections based of a paired reading
obtained with a polycarbonate filter (long-pass with cut-off at 400 nm). This
is a bit more sofisticated than simple subtracting the filter reading from
the mesurement as the effect of the filter itself on stray light is corrected
for.
}
\note{
The default \code{worker_fun} is just an example. Corrections are
specific to each individual spectrometer unit (not type or configuration)
and need to be written for each use case. The slit function tail correction
requires the characterization of the shape of the slit function by
measuring one or more laser beams at suitable wavelengths.

Currently \code{uvb_corrections_nb_files} allows processing of files
  written by OceanOptics' SpectraSuite software, from a simple protocol with
  no integration-time bracketing, with three measurements: a "light"
  measurement, a "filter" measurement using a polycarbonate filter and a dark
  measurement. Data should be raw counts, either corrected for detector
  non-linearity or not. All three spectra should be acquired using the
  same instrument settings to achieve good accuracy.
}
\author{
Algorithm for method "original" developed by Lasse Ylianttila. Other
  methods are modified from Ylianttila's method by Pedro J. Aphalo.
}
\references{
\url{http://www.r4photobiology.info}
}

