% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvb-corrections.R
\name{uvb_corrections}
\alias{filter_correction}
\alias{no_filter_correction}
\alias{slit_function_correction}
\alias{uvb_corrections}
\title{Apply filter-based stray-light correction}
\usage{
uvb_corrections(x, flt, dark, stray.light.method = "original",
  stray.light.wl = c(218.5, 228.5), flt.dark.wl = c(193, 209.5),
  flt.ref.wl = c(360, 379.5), inst.dark.pixs = c(1:4),
  worker.fun = ooacquire::maya_tail_correction, trim = 0, verbose = FALSE,
  ...)

slit_function_correction(x, worker.fun = ooacquire::maya_tail_correction,
  verbose = TRUE, ...)

filter_correction(x, flt, stray.light.method = "original",
  stray.light.wl = c(218.5, 228.5), flt.dark.wl = c(193, 209.5),
  flt.ref.wl = c(360, 379.5), trim = 0, verbose = FALSE)

no_filter_correction(x, stray.light.wl = c(218.5, 228.5),
  flt.dark.wl = c(193, 209.5), trim = 0, verbose = FALSE)
}
\arguments{
\item{x, flt, dark}{\code{cps_spct} objects.}

\item{stray.light.method}{Method variant used "original" (Ylianttila).}

\item{stray.light.wl}{numeric vector of length 2 giving the range of
wavelengths to use for the final stray light correction.}

\item{flt.dark.wl, flt.ref.wl}{numeric vectors of length 2 giving the ranges
of wavelengths to use for the "dark" and "illuminated" regions of the
array in the filter correction.}

\item{inst.dark.pixs}{numeric vector with indexes to array pixels that
are in full drakness by instrument design.}

\item{worker.fun}{function actually doing the correction on the w.lengths and
counts per second vectors.}

\item{trim}{a numeric value to be used as argument for mean}

\item{verbose}{Logical indicating the level of warnings wanted.}

\item{...}{additional params passed to worker.fun.}
}
\value{
A copy of \code{x} with the "cps" data replaced by the corrected
  ones.
}
\description{
Apply to a counts-per-second spectrum corrections based of a paired reading
obtained with a polycarbonate filter (long-pass with cut-off at 400 nm). This
is a bit more sofisticated than simple subtracting the filter reading from
the mesurement as the effect of the filter itself on stray light is corrected
for.
}
\note{
The default \code{worker.fun} is just an example. Corrections are
specific to each individual spectrometer unit (not type or configuration)
and need to be written for each use case. The slit function tail correction
requires the characterization of the shape of the slit function by
measuring one or more laser beams at suitable wavelengths.
}
\author{
Algorithm for method "original" developed by Lasse Ylianttila. Other
  methods are modified from Ylianttila's method by Pedro J. Aphalo.
}
\references{
\url{http://www.r4photobiology.info}
}

