% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example-raw-data.R
\docType{data}
\name{white_LED.raw_mspct}
\alias{white_LED.raw_mspct}
\title{Raw counts data for a lamp measurement.}
\format{
A raw_mspct
}
\usage{
white_LED.raw_mspct
}
\description{
A household Osram LED STAR CLASSIC A 60 lamp measured at short distance. The
spectrometer used was an Ocean Optics Maya2000 Pro.
}
\seealso{
Other objects containing example raw-counts data: 
\code{\link{blue_filter.raw_mspct}},
\code{\link{halogen.raw_mspct}},
\code{\link{red_filter.raw_mspct}},
\code{\link{white_LED_2min.raw_mspct}},
\code{\link{xenon_flash.raw_mspct}}
}
\concept{objects containing example raw-counts data}
\keyword{datasets}
